splitting simplify.py
splitting simplify.py into smaller files
