+++ sympy/core/tests/test_expand.py
@@ -1,6 +1,7 @@
 from sympy import (log, sqrt, Rational as R, Symbol, I, exp, pi, S,
     cos, sin, Mul, Pow, O)
-from sympy.simplify.simplify import expand_numer, expand
+from sympy.simplify.radsimp import expand_numer
+from sympy.simplify.simplify import expand
 from sympy.core.function import expand_multinomial, expand_power_base
 from sympy.core.compatibility import range
