+++ sympy/series/gruntz.py
@@ -121,7 +121,7 @@ from __future__ import print_function, division
 from sympy.core import Basic, S, oo, Symbol, I, Dummy, Wild, Mul
 from sympy.functions import log, exp
 from sympy.series.order import Order
-from sympy.simplify import powsimp
+from sympy.simplify.powsimp import powsimp
 from sympy import cacheit
 
 from sympy.core.compatibility import reduce
