+++ sympy/simplify/__init__.py
@@ -4,11 +4,10 @@ for example:
 the expression E**(pi*I) will be converted into -1
 the expression (x+x)**2 will be converted into 4*x**2
 """
-from .simplify import (collect, rcollect, radsimp, ratsimp, fraction,
-    simplify, trigsimp, powsimp, combsimp, hypersimp, hypersimilar, nsimplify,
-    logcombine, separatevars, numer, denom, powdenest, posify, polarify,
-    unpolarify, collect_const, signsimp, besselsimp, ratsimpmodprime,
-    exptrigsimp)
+from .simplify import (ratsimp,
+    simplify, combsimp, hypersimp, hypersimilar,
+    logcombine, separatevars, posify, polarify,
+    unpolarify, signsimp, ratsimpmodprime, bottom_up, nsimplify)
 
 from .fu import FU, fu
 
@@ -21,3 +20,11 @@ from .traversaltools import use
 from .epathtools import epath, EPath
 
 from .hyperexpand import hyperexpand
+
+from .radsimp import collect, rcollect, radsimp, collect_const, fraction, numer, denom
+
+from .trigsimp import trigsimp, exptrigsimp
+
+from .besselsimp import besselsimp
+
+from .powsimp import powsimp, powdenest
