+++ sympy/simplify/simplify.py
@@ -2,22 +2,18 @@ from __future__ import print_function, division
 
 from collections import defaultdict
 
-from sympy import SYMPY_DEBUG
-
 from sympy.core import (Basic, S, Add, Mul, Pow,
-    Derivative, Wild, Symbol, sympify, expand, expand_mul, expand_func,
+    Symbol, sympify, expand, expand_mul, expand_func,
     Function, Dummy, Expr, factor_terms,
-    FunctionClass, expand_power_base, symbols, igcd,
+    FunctionClass, symbols,
     expand_power_exp)
 from sympy.core.add import _unevaluated_Add
-from sympy.core.cache import cacheit
 from sympy.core.compatibility import (combinations_with_replacement, iterable,
     reduce, default_sort_key,
     ordered, range, as_int)
-from sympy.core.exprtools import Factors, gcd_terms
 from sympy.core.numbers import Float, I, pi, Rational, Integer
 from sympy.core.function import expand_log, count_ops, _mexpand
-from sympy.core.mul import _keep_coeff, prod, _unevaluated_Mul
+from sympy.core.mul import prod, _unevaluated_Mul
 from sympy.core.rules import Transform
 from sympy.core.evaluate import global_evaluate
 from sympy.functions import (
@@ -29,552 +25,28 @@ from sympy.functions.elementary.exponential import ExpBase
 from sympy.functions.elementary.hyperbolic import HyperbolicFunction
 from sympy.functions.elementary.integers import ceiling
 from sympy.functions.elementary.trigonometric import TrigonometricFunction
-from sympy.functions.special.bessel import (
-    besselj, bessely, besseli, besselk, jn)
-
-from sympy.strategies.core import identity
-from sympy.strategies.tree import greedy
 
 from sympy.utilities.iterables import has_variety, sift, uniq
 
 from sympy.utilities.misc import debug
 from sympy.utilities.timeutils import timethis
 
-from sympy.simplify.cse_main import cse
+
+from sympy.simplify.radsimp import radsimp, collect_sqrt, fraction
+from sympy.simplify.trigsimp import trigsimp, exptrigsimp
+from sympy.simplify.besselsimp import besselsimp
+from sympy.simplify.powsimp import powsimp
 from sympy.simplify.cse_opts import sub_pre, sub_post
 from sympy.simplify.sqrtdenest import sqrtdenest
-from sympy.ntheory.factor_ import multiplicity
 
 from sympy.polys import (Poly, together, reduced, cancel, factor,
     ComputationFailed, lcm, gcd, parallel_poly_from_expr)
-from sympy.polys.domains import ZZ
 from sympy.polys.monomials import Monomial, monomial_div
 from sympy.polys.polyerrors import PolificationFailed, DomainError
-from sympy.polys.polytools import groebner
+
 
 import mpmath
 
-
-
-def fraction(expr, exact=False):
-    """Returns a pair with expression's numerator and denominator.
-       If the given expression is not a fraction then this function
-       will return the tuple (expr, 1).
-
-       This function will not make any attempt to simplify nested
-       fractions or to do any term rewriting at all.
-
-       If only one of the numerator/denominator pair is needed then
-       use numer(expr) or denom(expr) functions respectively.
-
-       >>> from sympy import fraction, Rational, Symbol
-       >>> from sympy.abc import x, y
-
-       >>> fraction(x/y)
-       (x, y)
-       >>> fraction(x)
-       (x, 1)
-
-       >>> fraction(1/y**2)
-       (1, y**2)
-
-       >>> fraction(x*y/2)
-       (x*y, 2)
-       >>> fraction(Rational(1, 2))
-       (1, 2)
-
-       This function will also work fine with assumptions:
-
-       >>> k = Symbol('k', negative=True)
-       >>> fraction(x * y**k)
-       (x, y**(-k))
-
-       If we know nothing about sign of some exponent and 'exact'
-       flag is unset, then structure this exponent's structure will
-       be analyzed and pretty fraction will be returned:
-
-       >>> from sympy import exp
-       >>> fraction(2*x**(-y))
-       (2, x**y)
-
-       >>> fraction(exp(-x))
-       (1, exp(x))
-
-       >>> fraction(exp(-x), exact=True)
-       (exp(-x), 1)
-
-    """
-    expr = sympify(expr)
-
-    numer, denom = [], []
-
-    for term in Mul.make_args(expr):
-        if term.is_commutative and (term.is_Pow or term.func is exp):
-            b, ex = term.as_base_exp()
-            if ex.is_negative:
-                if ex is S.NegativeOne:
-                    denom.append(b)
-                else:
-                    denom.append(Pow(b, -ex))
-            elif ex.is_positive:
-                numer.append(term)
-            elif not exact and ex.is_Mul:
-                n, d = term.as_numer_denom()
-                numer.append(n)
-                denom.append(d)
-            else:
-                numer.append(term)
-        elif term.is_Rational:
-            n, d = term.as_numer_denom()
-            numer.append(n)
-            denom.append(d)
-        else:
-            numer.append(term)
-
-    return Mul(*numer), Mul(*denom)
-
-
-def numer(expr):
-    return fraction(expr)[0]
-
-
-def denom(expr):
-    return fraction(expr)[1]
-
-
-def fraction_expand(expr, **hints):
-    return expr.expand(frac=True, **hints)
-
-
-def numer_expand(expr, **hints):
-    a, b = fraction(expr)
-    return a.expand(numer=True, **hints) / b
-
-
-def denom_expand(expr, **hints):
-    a, b = fraction(expr)
-    return a / b.expand(denom=True, **hints)
-
-expand_numer = numer_expand
-expand_denom = denom_expand
-expand_fraction = fraction_expand
-
-
-def collect(expr, syms, func=None, evaluate=None, exact=False, distribute_order_term=True):
-    """
-    Collect additive terms of an expression.
-
-    This function collects additive terms of an expression with respect
-    to a list of expression up to powers with rational exponents. By the
-    term symbol here are meant arbitrary expressions, which can contain
-    powers, products, sums etc. In other words symbol is a pattern which
-    will be searched for in the expression's terms.
-
-    The input expression is not expanded by :func:`collect`, so user is
-    expected to provide an expression is an appropriate form. This makes
-    :func:`collect` more predictable as there is no magic happening behind the
-    scenes. However, it is important to note, that powers of products are
-    converted to products of powers using the :func:`expand_power_base`
-    function.
-
-    There are two possible types of output. First, if ``evaluate`` flag is
-    set, this function will return an expression with collected terms or
-    else it will return a dictionary with expressions up to rational powers
-    as keys and collected coefficients as values.
-
-    Examples
-    ========
-
-    >>> from sympy import S, collect, expand, factor, Wild
-    >>> from sympy.abc import a, b, c, x, y, z
-
-    This function can collect symbolic coefficients in polynomials or
-    rational expressions. It will manage to find all integer or rational
-    powers of collection variable::
-
-        >>> collect(a*x**2 + b*x**2 + a*x - b*x + c, x)
-        c + x**2*(a + b) + x*(a - b)
-
-    The same result can be achieved in dictionary form::
-
-        >>> d = collect(a*x**2 + b*x**2 + a*x - b*x + c, x, evaluate=False)
-        >>> d[x**2]
-        a + b
-        >>> d[x]
-        a - b
-        >>> d[S.One]
-        c
-
-    You can also work with multivariate polynomials. However, remember that
-    this function is greedy so it will care only about a single symbol at time,
-    in specification order::
-
-        >>> collect(x**2 + y*x**2 + x*y + y + a*y, [x, y])
-        x**2*(y + 1) + x*y + y*(a + 1)
-
-    Also more complicated expressions can be used as patterns::
-
-        >>> from sympy import sin, log
-        >>> collect(a*sin(2*x) + b*sin(2*x), sin(2*x))
-        (a + b)*sin(2*x)
-
-        >>> collect(a*x*log(x) + b*(x*log(x)), x*log(x))
-        x*(a + b)*log(x)
-
-    You can use wildcards in the pattern::
-
-        >>> w = Wild('w1')
-        >>> collect(a*x**y - b*x**y, w**y)
-        x**y*(a - b)
-
-    It is also possible to work with symbolic powers, although it has more
-    complicated behavior, because in this case power's base and symbolic part
-    of the exponent are treated as a single symbol::
-
-        >>> collect(a*x**c + b*x**c, x)
-        a*x**c + b*x**c
-        >>> collect(a*x**c + b*x**c, x**c)
-        x**c*(a + b)
-
-    However if you incorporate rationals to the exponents, then you will get
-    well known behavior::
-
-        >>> collect(a*x**(2*c) + b*x**(2*c), x**c)
-        x**(2*c)*(a + b)
-
-    Note also that all previously stated facts about :func:`collect` function
-    apply to the exponential function, so you can get::
-
-        >>> from sympy import exp
-        >>> collect(a*exp(2*x) + b*exp(2*x), exp(x))
-        (a + b)*exp(2*x)
-
-    If you are interested only in collecting specific powers of some symbols
-    then set ``exact`` flag in arguments::
-
-        >>> collect(a*x**7 + b*x**7, x, exact=True)
-        a*x**7 + b*x**7
-        >>> collect(a*x**7 + b*x**7, x**7, exact=True)
-        x**7*(a + b)
-
-    You can also apply this function to differential equations, where
-    derivatives of arbitrary order can be collected. Note that if you
-    collect with respect to a function or a derivative of a function, all
-    derivatives of that function will also be collected. Use
-    ``exact=True`` to prevent this from happening::
-
-        >>> from sympy import Derivative as D, collect, Function
-        >>> f = Function('f') (x)
-
-        >>> collect(a*D(f,x) + b*D(f,x), D(f,x))
-        (a + b)*Derivative(f(x), x)
-
-        >>> collect(a*D(D(f,x),x) + b*D(D(f,x),x), f)
-        (a + b)*Derivative(f(x), x, x)
-
-        >>> collect(a*D(D(f,x),x) + b*D(D(f,x),x), D(f,x), exact=True)
-        a*Derivative(f(x), x, x) + b*Derivative(f(x), x, x)
-
-        >>> collect(a*D(f,x) + b*D(f,x) + a*f + b*f, f)
-        (a + b)*f(x) + (a + b)*Derivative(f(x), x)
-
-    Or you can even match both derivative order and exponent at the same time::
-
-        >>> collect(a*D(D(f,x),x)**2 + b*D(D(f,x),x)**2, D(f,x))
-        (a + b)*Derivative(f(x), x, x)**2
-
-    Finally, you can apply a function to each of the collected coefficients.
-    For example you can factorize symbolic coefficients of polynomial::
-
-        >>> f = expand((x + a + 1)**3)
-
-        >>> collect(f, x, factor)
-        x**3 + 3*x**2*(a + 1) + 3*x*(a + 1)**2 + (a + 1)**3
-
-    .. note:: Arguments are expected to be in expanded form, so you might have
-              to call :func:`expand` prior to calling this function.
-
-    See Also
-    ========
-    collect_const, collect_sqrt, rcollect
-    """
-    if evaluate is None:
-        evaluate = global_evaluate[0]
-
-    def make_expression(terms):
-        product = []
-
-        for term, rat, sym, deriv in terms:
-            if deriv is not None:
-                var, order = deriv
-
-                while order > 0:
-                    term, order = Derivative(term, var), order - 1
-
-            if sym is None:
-                if rat is S.One:
-                    product.append(term)
-                else:
-                    product.append(Pow(term, rat))
-            else:
-                product.append(Pow(term, rat*sym))
-
-        return Mul(*product)
-
-    def parse_derivative(deriv):
-        # scan derivatives tower in the input expression and return
-        # underlying function and maximal differentiation order
-        expr, sym, order = deriv.expr, deriv.variables[0], 1
-
-        for s in deriv.variables[1:]:
-            if s == sym:
-                order += 1
-            else:
-                raise NotImplementedError(
-                    'Improve MV Derivative support in collect')
-
-        while isinstance(expr, Derivative):
-            s0 = expr.variables[0]
-
-            for s in expr.variables:
-                if s != s0:
-                    raise NotImplementedError(
-                        'Improve MV Derivative support in collect')
-
-            if s0 == sym:
-                expr, order = expr.expr, order + len(expr.variables)
-            else:
-                break
-
-        return expr, (sym, Rational(order))
-
-    def parse_term(expr):
-        """Parses expression expr and outputs tuple (sexpr, rat_expo,
-        sym_expo, deriv)
-        where:
-         - sexpr is the base expression
-         - rat_expo is the rational exponent that sexpr is raised to
-         - sym_expo is the symbolic exponent that sexpr is raised to
-         - deriv contains the derivatives the the expression
-
-         for example, the output of x would be (x, 1, None, None)
-         the output of 2**x would be (2, 1, x, None)
-        """
-        rat_expo, sym_expo = S.One, None
-        sexpr, deriv = expr, None
-
-        if expr.is_Pow:
-            if isinstance(expr.base, Derivative):
-                sexpr, deriv = parse_derivative(expr.base)
-            else:
-                sexpr = expr.base
-
-            if expr.exp.is_Number:
-                rat_expo = expr.exp
-            else:
-                coeff, tail = expr.exp.as_coeff_Mul()
-
-                if coeff.is_Number:
-                    rat_expo, sym_expo = coeff, tail
-                else:
-                    sym_expo = expr.exp
-        elif expr.func is exp:
-            arg = expr.args[0]
-            if arg.is_Rational:
-                sexpr, rat_expo = S.Exp1, arg
-            elif arg.is_Mul:
-                coeff, tail = arg.as_coeff_Mul(rational=True)
-                sexpr, rat_expo = exp(tail), coeff
-        elif isinstance(expr, Derivative):
-            sexpr, deriv = parse_derivative(expr)
-
-        return sexpr, rat_expo, sym_expo, deriv
-
-    def parse_expression(terms, pattern):
-        """Parse terms searching for a pattern.
-        terms is a list of tuples as returned by parse_terms;
-        pattern is an expression treated as a product of factors
-        """
-        pattern = Mul.make_args(pattern)
-
-        if len(terms) < len(pattern):
-            # pattern is longer than matched product
-            # so no chance for positive parsing result
-            return None
-        else:
-            pattern = [parse_term(elem) for elem in pattern]
-
-            terms = terms[:]  # need a copy
-            elems, common_expo, has_deriv = [], None, False
-
-            for elem, e_rat, e_sym, e_ord in pattern:
-
-                if elem.is_Number and e_rat == 1 and e_sym is None:
-                    # a constant is a match for everything
-                    continue
-
-                for j in range(len(terms)):
-                    if terms[j] is None:
-                        continue
-
-                    term, t_rat, t_sym, t_ord = terms[j]
-
-                    # keeping track of whether one of the terms had
-                    # a derivative or not as this will require rebuilding
-                    # the expression later
-                    if t_ord is not None:
-                        has_deriv = True
-
-                    if (term.match(elem) is not None and
-                            (t_sym == e_sym or t_sym is not None and
-                            e_sym is not None and
-                            t_sym.match(e_sym) is not None)):
-                        if exact is False:
-                            # we don't have to be exact so find common exponent
-                            # for both expression's term and pattern's element
-                            expo = t_rat / e_rat
-
-                            if common_expo is None:
-                                # first time
-                                common_expo = expo
-                            else:
-                                # common exponent was negotiated before so
-                                # there is no chance for a pattern match unless
-                                # common and current exponents are equal
-                                if common_expo != expo:
-                                    common_expo = 1
-                        else:
-                            # we ought to be exact so all fields of
-                            # interest must match in every details
-                            if e_rat != t_rat or e_ord != t_ord:
-                                continue
-
-                        # found common term so remove it from the expression
-                        # and try to match next element in the pattern
-                        elems.append(terms[j])
-                        terms[j] = None
-
-                        break
-
-                else:
-                    # pattern element not found
-                    return None
-
-            return [_f for _f in terms if _f], elems, common_expo, has_deriv
-
-    if evaluate:
-        if expr.is_Mul:
-            return expr.func(*[
-                collect(term, syms, func, True, exact, distribute_order_term)
-                for term in expr.args])
-        elif expr.is_Pow:
-            b = collect(
-                expr.base, syms, func, True, exact, distribute_order_term)
-            return Pow(b, expr.exp)
-
-    if iterable(syms):
-        syms = [expand_power_base(i, deep=False) for i in syms]
-    else:
-        syms = [expand_power_base(syms, deep=False)]
-
-    expr = sympify(expr)
-    order_term = None
-
-    if distribute_order_term:
-        order_term = expr.getO()
-
-        if order_term is not None:
-            if order_term.has(*syms):
-                order_term = None
-            else:
-                expr = expr.removeO()
-
-    summa = [expand_power_base(i, deep=False) for i in Add.make_args(expr)]
-
-    collected, disliked = defaultdict(list), S.Zero
-    for product in summa:
-        terms = [parse_term(i) for i in Mul.make_args(product)]
-
-        for symbol in syms:
-            if SYMPY_DEBUG:
-                print("DEBUG: parsing of expression %s with symbol %s " % (
-                    str(terms), str(symbol))
-                )
-
-            result = parse_expression(terms, symbol)
-
-            if SYMPY_DEBUG:
-                print("DEBUG: returned %s" % str(result))
-
-            if result is not None:
-                terms, elems, common_expo, has_deriv = result
-
-                # when there was derivative in current pattern we
-                # will need to rebuild its expression from scratch
-                if not has_deriv:
-                    index = 1
-                    for elem in elems:
-                        e = elem[1]
-                        if elem[2] is not None:
-                            e *= elem[2]
-                        index *= Pow(elem[0], e)
-                else:
-                    index = make_expression(elems)
-                terms = expand_power_base(make_expression(terms), deep=False)
-                index = expand_power_base(index, deep=False)
-                collected[index].append(terms)
-                break
-        else:
-            # none of the patterns matched
-            disliked += product
-    # add terms now for each key
-    collected = dict([(k, Add(*v)) for k, v in collected.items()])
-
-    if disliked is not S.Zero:
-        collected[S.One] = disliked
-
-    if order_term is not None:
-        for key, val in collected.items():
-            collected[key] = val + order_term
-
-    if func is not None:
-        collected = dict(
-            [(key, func(val)) for key, val in collected.items()])
-
-    if evaluate:
-        return Add(*[key*val for key, val in collected.items()])
-    else:
-        return collected
-
-
-def rcollect(expr, *vars):
-    """
-    Recursively collect sums in an expression.
-
-    Examples
-    ========
-
-    >>> from sympy.simplify import rcollect
-    >>> from sympy.abc import x, y
-
-    >>> expr = (x**2*y + x*y + x + y)/(x + y)
-
-    >>> rcollect(expr, y)
-    (x + y*(x**2 + x + 1))/(x + y)
-
-    See Also
-    ========
-    collect, collect_const, collect_sqrt
-    """
-    if expr.is_Atom or not expr.has(*vars):
-        return expr
-    else:
-        expr = expr.__class__(*[rcollect(arg, *vars) for arg in expr.args])
-
-        if expr.is_Add:
-            return collect(expr, vars)
-        else:
-            return expr
 
 
 def separatevars(expr, symbols=[], dict=False, force=False):
@@ -943,1155 +415,6 @@ def ratsimpmodprime(expr, G, *gens, **args):
     return (c*r.q)/(d*r.p)
 
 
-def trigsimp_groebner(expr, hints=[], quick=False, order="grlex",
-                      polynomial=False):
-    """
-    Simplify trigonometric expressions using a groebner basis algorithm.
-
-    This routine takes a fraction involving trigonometric or hyperbolic
-    expressions, and tries to simplify it. The primary metric is the
-    total degree. Some attempts are made to choose the simplest possible
-    expression of the minimal degree, but this is non-rigorous, and also
-    very slow (see the ``quick=True`` option).
-
-    If ``polynomial`` is set to True, instead of simplifying numerator and
-    denominator together, this function just brings numerator and denominator
-    into a canonical form. This is much faster, but has potentially worse
-    results. However, if the input is a polynomial, then the result is
-    guaranteed to be an equivalent polynomial of minimal degree.
-
-    The most important option is hints. Its entries can be any of the
-    following:
-
-    - a natural number
-    - a function
-    - an iterable of the form (func, var1, var2, ...)
-    - anything else, interpreted as a generator
-
-    A number is used to indicate that the search space should be increased.
-    A function is used to indicate that said function is likely to occur in a
-    simplified expression.
-    An iterable is used indicate that func(var1 + var2 + ...) is likely to
-    occur in a simplified .
-    An additional generator also indicates that it is likely to occur.
-    (See examples below).
-
-    This routine carries out various computationally intensive algorithms.
-    The option ``quick=True`` can be used to suppress one particularly slow
-    step (at the expense of potentially more complicated results, but never at
-    the expense of increased total degree).
-
-    Examples
-    ========
-
-    >>> from sympy.abc import x, y
-    >>> from sympy import sin, tan, cos, sinh, cosh, tanh
-    >>> from sympy.simplify.simplify import trigsimp_groebner
-
-    Suppose you want to simplify ``sin(x)*cos(x)``. Naively, nothing happens:
-
-    >>> ex = sin(x)*cos(x)
-    >>> trigsimp_groebner(ex)
-    sin(x)*cos(x)
-
-    This is because ``trigsimp_groebner`` only looks for a simplification
-    involving just ``sin(x)`` and ``cos(x)``. You can tell it to also try
-    ``2*x`` by passing ``hints=[2]``:
-
-    >>> trigsimp_groebner(ex, hints=[2])
-    sin(2*x)/2
-    >>> trigsimp_groebner(sin(x)**2 - cos(x)**2, hints=[2])
-    -cos(2*x)
-
-    Increasing the search space this way can quickly become expensive. A much
-    faster way is to give a specific expression that is likely to occur:
-
-    >>> trigsimp_groebner(ex, hints=[sin(2*x)])
-    sin(2*x)/2
-
-    Hyperbolic expressions are similarly supported:
-
-    >>> trigsimp_groebner(sinh(2*x)/sinh(x))
-    2*cosh(x)
-
-    Note how no hints had to be passed, since the expression already involved
-    ``2*x``.
-
-    The tangent function is also supported. You can either pass ``tan`` in the
-    hints, to indicate that than should be tried whenever cosine or sine are,
-    or you can pass a specific generator:
-
-    >>> trigsimp_groebner(sin(x)/cos(x), hints=[tan])
-    tan(x)
-    >>> trigsimp_groebner(sinh(x)/cosh(x), hints=[tanh(x)])
-    tanh(x)
-
-    Finally, you can use the iterable form to suggest that angle sum formulae
-    should be tried:
-
-    >>> ex = (tan(x) + tan(y))/(1 - tan(x)*tan(y))
-    >>> trigsimp_groebner(ex, hints=[(tan, x, y)])
-    tan(x + y)
-    """
-    # TODO
-    #  - preprocess by replacing everything by funcs we can handle
-    # - optionally use cot instead of tan
-    # - more intelligent hinting.
-    #     For example, if the ideal is small, and we have sin(x), sin(y),
-    #     add sin(x + y) automatically... ?
-    # - algebraic numbers ...
-    # - expressions of lowest degree are not distinguished properly
-    #   e.g. 1 - sin(x)**2
-    # - we could try to order the generators intelligently, so as to influence
-    #   which monomials appear in the quotient basis
-
-    # THEORY
-    # ------
-    # Ratsimpmodprime above can be used to "simplify" a rational function
-    # modulo a prime ideal. "Simplify" mainly means finding an equivalent
-    # expression of lower total degree.
-    #
-    # We intend to use this to simplify trigonometric functions. To do that,
-    # we need to decide (a) which ring to use, and (b) modulo which ideal to
-    # simplify. In practice, (a) means settling on a list of "generators"
-    # a, b, c, ..., such that the fraction we want to simplify is a rational
-    # function in a, b, c, ..., with coefficients in ZZ (integers).
-    # (2) means that we have to decide what relations to impose on the
-    # generators. There are two practical problems:
-    #   (1) The ideal has to be *prime* (a technical term).
-    #   (2) The relations have to be polynomials in the generators.
-    #
-    # We typically have two kinds of generators:
-    # - trigonometric expressions, like sin(x), cos(5*x), etc
-    # - "everything else", like gamma(x), pi, etc.
-    #
-    # Since this function is trigsimp, we will concentrate on what to do with
-    # trigonometric expressions. We can also simplify hyperbolic expressions,
-    # but the extensions should be clear.
-    #
-    # One crucial point is that all *other* generators really should behave
-    # like indeterminates. In particular if (say) "I" is one of them, then
-    # in fact I**2 + 1 = 0 and we may and will compute non-sensical
-    # expressions. However, we can work with a dummy and add the relation
-    # I**2 + 1 = 0 to our ideal, then substitute back in the end.
-    #
-    # Now regarding trigonometric generators. We split them into groups,
-    # according to the argument of the trigonometric functions. We want to
-    # organise this in such a way that most trigonometric identities apply in
-    # the same group. For example, given sin(x), cos(2*x) and cos(y), we would
-    # group as [sin(x), cos(2*x)] and [cos(y)].
-    #
-    # Our prime ideal will be built in three steps:
-    # (1) For each group, compute a "geometrically prime" ideal of relations.
-    #     Geometrically prime means that it generates a prime ideal in
-    #     CC[gens], not just ZZ[gens].
-    # (2) Take the union of all the generators of the ideals for all groups.
-    #     By the geometric primality condition, this is still prime.
-    # (3) Add further inter-group relations which preserve primality.
-    #
-    # Step (1) works as follows. We will isolate common factors in the
-    # argument, so that all our generators are of the form sin(n*x), cos(n*x)
-    # or tan(n*x), with n an integer. Suppose first there are no tan terms.
-    # The ideal [sin(x)**2 + cos(x)**2 - 1] is geometrically prime, since
-    # X**2 + Y**2 - 1 is irreducible over CC.
-    # Now, if we have a generator sin(n*x), than we can, using trig identities,
-    # express sin(n*x) as a polynomial in sin(x) and cos(x). We can add this
-    # relation to the ideal, preserving geometric primality, since the quotient
-    # ring is unchanged.
-    # Thus we have treated all sin and cos terms.
-    # For tan(n*x), we add a relation tan(n*x)*cos(n*x) - sin(n*x) = 0.
-    # (This requires of course that we already have relations for cos(n*x) and
-    # sin(n*x).) It is not obvious, but it seems that this preserves geometric
-    # primality.
-    # XXX A real proof would be nice. HELP!
-    #     Sketch that <S**2 + C**2 - 1, C*T - S> is a prime ideal of
-    #     CC[S, C, T]:
-    #     - it suffices to show that the projective closure in CP**3 is
-    #       irreducible
-    #     - using the half-angle substitutions, we can express sin(x), tan(x),
-    #       cos(x) as rational functions in tan(x/2)
-    #     - from this, we get a rational map from CP**1 to our curve
-    #     - this is a morphism, hence the curve is prime
-    #
-    # Step (2) is trivial.
-    #
-    # Step (3) works by adding selected relations of the form
-    # sin(x + y) - sin(x)*cos(y) - sin(y)*cos(x), etc. Geometric primality is
-    # preserved by the same argument as before.
-
-    def parse_hints(hints):
-        """Split hints into (n, funcs, iterables, gens)."""
-        n = 1
-        funcs, iterables, gens = [], [], []
-        for e in hints:
-            if isinstance(e, (int, Integer)):
-                n = e
-            elif isinstance(e, FunctionClass):
-                funcs.append(e)
-            elif iterable(e):
-                iterables.append((e[0], e[1:]))
-                # XXX sin(x+2y)?
-                # Note: we go through polys so e.g.
-                # sin(-x) -> -sin(x) -> sin(x)
-                gens.extend(parallel_poly_from_expr(
-                    [e[0](x) for x in e[1:]] + [e[0](Add(*e[1:]))])[1].gens)
-            else:
-                gens.append(e)
-        return n, funcs, iterables, gens
-
-    def build_ideal(x, terms):
-        """
-        Build generators for our ideal. Terms is an iterable with elements of
-        the form (fn, coeff), indicating that we have a generator fn(coeff*x).
-
-        If any of the terms is trigonometric, sin(x) and cos(x) are guaranteed
-        to appear in terms. Similarly for hyperbolic functions. For tan(n*x),
-        sin(n*x) and cos(n*x) are guaranteed.
-        """
-        gens = []
-        I = []
-        y = Dummy('y')
-        for fn, coeff in terms:
-            for c, s, t, rel in (
-                    [cos, sin, tan, cos(x)**2 + sin(x)**2 - 1],
-                    [cosh, sinh, tanh, cosh(x)**2 - sinh(x)**2 - 1]):
-                if coeff == 1 and fn in [c, s]:
-                    I.append(rel)
-                elif fn == t:
-                    I.append(t(coeff*x)*c(coeff*x) - s(coeff*x))
-                elif fn in [c, s]:
-                    cn = fn(coeff*y).expand(trig=True).subs(y, x)
-                    I.append(fn(coeff*x) - cn)
-        return list(set(I))
-
-    def analyse_gens(gens, hints):
-        """
-        Analyse the generators ``gens``, using the hints ``hints``.
-
-        The meaning of ``hints`` is described in the main docstring.
-        Return a new list of generators, and also the ideal we should
-        work with.
-        """
-        # First parse the hints
-        n, funcs, iterables, extragens = parse_hints(hints)
-        debug('n=%s' % n, 'funcs:', funcs, 'iterables:',
-              iterables, 'extragens:', extragens)
-
-        # We just add the extragens to gens and analyse them as before
-        gens = list(gens)
-        gens.extend(extragens)
-
-        # remove duplicates
-        funcs = list(set(funcs))
-        iterables = list(set(iterables))
-        gens = list(set(gens))
-
-        # all the functions we can do anything with
-        allfuncs = set([sin, cos, tan, sinh, cosh, tanh])
-        # sin(3*x) -> ((3, x), sin)
-        trigterms = [(g.args[0].as_coeff_mul(), g.func) for g in gens
-                     if g.func in allfuncs]
-        # Our list of new generators - start with anything that we cannot
-        # work with (i.e. is not a trigonometric term)
-        freegens = [g for g in gens if g.func not in allfuncs]
-        newgens = []
-        trigdict = {}
-        for (coeff, var), fn in trigterms:
-            trigdict.setdefault(var, []).append((coeff, fn))
-        res = [] # the ideal
-
-        for key, val in trigdict.items():
-            # We have now assembeled a dictionary. Its keys are common
-            # arguments in trigonometric expressions, and values are lists of
-            # pairs (fn, coeff). x0, (fn, coeff) in trigdict means that we
-            # need to deal with fn(coeff*x0). We take the rational gcd of the
-            # coeffs, call it ``gcd``. We then use x = x0/gcd as "base symbol",
-            # all other arguments are integral multiples thereof.
-            # We will build an ideal which works with sin(x), cos(x).
-            # If hint tan is provided, also work with tan(x). Moreover, if
-            # n > 1, also work with sin(k*x) for k <= n, and similarly for cos
-            # (and tan if the hint is provided). Finally, any generators which
-            # the ideal does not work with but we need to accomodate (either
-            # because it was in expr or because it was provided as a hint)
-            # we also build into the ideal.
-            # This selection process is expressed in the list ``terms``.
-            # build_ideal then generates the actual relations in our ideal,
-            # from this list.
-            fns = [x[1] for x in val]
-            val = [x[0] for x in val]
-            gcd = reduce(igcd, val)
-            terms = [(fn, v/gcd) for (fn, v) in zip(fns, val)]
-            fs = set(funcs + fns)
-            for c, s, t in ([cos, sin, tan], [cosh, sinh, tanh]):
-                if any(x in fs for x in (c, s, t)):
-                    fs.add(c)
-                    fs.add(s)
-            for fn in fs:
-                for k in range(1, n + 1):
-                    terms.append((fn, k))
-            extra = []
-            for fn, v in terms:
-                if fn == tan:
-                    extra.append((sin, v))
-                    extra.append((cos, v))
-                if fn in [sin, cos] and tan in fs:
-                    extra.append((tan, v))
-                if fn == tanh:
-                    extra.append((sinh, v))
-                    extra.append((cosh, v))
-                if fn in [sinh, cosh] and tanh in fs:
-                    extra.append((tanh, v))
-            terms.extend(extra)
-            x = gcd*Mul(*key)
-            r = build_ideal(x, terms)
-            res.extend(r)
-            newgens.extend(set(fn(v*x) for fn, v in terms))
-
-        # Add generators for compound expressions from iterables
-        for fn, args in iterables:
-            if fn == tan:
-                # Tan expressions are recovered from sin and cos.
-                iterables.extend([(sin, args), (cos, args)])
-            elif fn == tanh:
-                # Tanh expressions are recovered from sihn and cosh.
-                iterables.extend([(sinh, args), (cosh, args)])
-            else:
-                dummys = symbols('d:%i' % len(args), cls=Dummy)
-                expr = fn( Add(*dummys)).expand(trig=True).subs(list(zip(dummys, args)))
-                res.append(fn(Add(*args)) - expr)
-
-        if myI in gens:
-            res.append(myI**2 + 1)
-            freegens.remove(myI)
-            newgens.append(myI)
-
-        return res, freegens, newgens
-
-    myI = Dummy('I')
-    expr = expr.subs(S.ImaginaryUnit, myI)
-    subs = [(myI, S.ImaginaryUnit)]
-
-    num, denom = cancel(expr).as_numer_denom()
-    try:
-        (pnum, pdenom), opt = parallel_poly_from_expr([num, denom])
-    except PolificationFailed:
-        return expr
-    debug('initial gens:', opt.gens)
-    ideal, freegens, gens = analyse_gens(opt.gens, hints)
-    debug('ideal:', ideal)
-    debug('new gens:', gens, " -- len", len(gens))
-    debug('free gens:', freegens, " -- len", len(gens))
-    # NOTE we force the domain to be ZZ to stop polys from injecting generators
-    #      (which is usually a sign of a bug in the way we build the ideal)
-    if not gens:
-        return expr
-    G = groebner(ideal, order=order, gens=gens, domain=ZZ)
-    debug('groebner basis:', list(G), " -- len", len(G))
-
-    # If our fraction is a polynomial in the free generators, simplify all
-    # coefficients separately:
-    if freegens and pdenom.has_only_gens(*set(gens).intersection(pdenom.gens)):
-        num = Poly(num, gens=gens+freegens).eject(*gens)
-        res = []
-        for monom, coeff in num.terms():
-            ourgens = set(parallel_poly_from_expr([coeff, denom])[1].gens)
-            # We compute the transitive closure of all generators that can
-            # be reached from our generators through relations in the ideal.
-            changed = True
-            while changed:
-                changed = False
-                for p in ideal:
-                    p = Poly(p)
-                    if not ourgens.issuperset(p.gens) and \
-                       not p.has_only_gens(*set(p.gens).difference(ourgens)):
-                        changed = True
-                        ourgens.update(p.exclude().gens)
-            # NOTE preserve order!
-            realgens = [x for x in gens if x in ourgens]
-            # The generators of the ideal have now been (implicitely) split
-            # into two groups: those involving ourgens and those that don't.
-            # Since we took the transitive closure above, these two groups
-            # live in subgrings generated by a *disjoint* set of variables.
-            # Any sensible groebner basis algorithm will preserve this disjoint
-            # structure (i.e. the elements of the groebner basis can be split
-            # similarly), and and the two subsets of the groebner basis then
-            # form groebner bases by themselves. (For the smaller generating
-            # sets, of course.)
-            ourG = [g.as_expr() for g in G.polys if
-                    g.has_only_gens(*ourgens.intersection(g.gens))]
-            res.append(Mul(*[a**b for a, b in zip(freegens, monom)]) * \
-                       ratsimpmodprime(coeff/denom, ourG, order=order,
-                                       gens=realgens, quick=quick, domain=ZZ,
-                                       polynomial=polynomial).subs(subs))
-        return Add(*res)
-        # NOTE The following is simpler and has less assumptions on the
-        #      groebner basis algorithm. If the above turns out to be broken,
-        #      use this.
-        return Add(*[Mul(*[a**b for a, b in zip(freegens, monom)]) * \
-                     ratsimpmodprime(coeff/denom, list(G), order=order,
-                                     gens=gens, quick=quick, domain=ZZ)
-                     for monom, coeff in num.terms()])
-    else:
-        return ratsimpmodprime(
-            expr, list(G), order=order, gens=freegens+gens,
-            quick=quick, domain=ZZ, polynomial=polynomial).subs(subs)
-
-
-_trigs = (TrigonometricFunction, HyperbolicFunction)
-
-
-def trigsimp(expr, **opts):
-    """
-    reduces expression by using known trig identities
-
-    Notes
-    =====
-
-    method:
-    - Determine the method to use. Valid choices are 'matching' (default),
-    'groebner', 'combined', and 'fu'. If 'matching', simplify the
-    expression recursively by targeting common patterns. If 'groebner', apply
-    an experimental groebner basis algorithm. In this case further options
-    are forwarded to ``trigsimp_groebner``, please refer to its docstring.
-    If 'combined', first run the groebner basis algorithm with small
-    default parameters, then run the 'matching' algorithm. 'fu' runs the
-    collection of trigonometric transformations described by Fu, et al.
-    (see the `fu` docstring).
-
-
-    Examples
-    ========
-
-    >>> from sympy import trigsimp, sin, cos, log
-    >>> from sympy.abc import x, y
-    >>> e = 2*sin(x)**2 + 2*cos(x)**2
-    >>> trigsimp(e)
-    2
-
-    Simplification occurs wherever trigonometric functions are located.
-
-    >>> trigsimp(log(e))
-    log(2)
-
-    Using `method="groebner"` (or `"combined"`) might lead to greater
-    simplification.
-
-    The old trigsimp routine can be accessed as with method 'old'.
-
-    >>> from sympy import coth, tanh
-    >>> t = 3*tanh(x)**7 - 2/coth(x)**7
-    >>> trigsimp(t, method='old') == t
-    True
-    >>> trigsimp(t)
-    tanh(x)**7
-
-    """
-    from sympy.simplify.fu import fu
-
-    expr = sympify(expr)
-
-    try:
-        return expr._eval_trigsimp(**opts)
-    except AttributeError:
-        pass
-
-    old = opts.pop('old', False)
-    if not old:
-        opts.pop('deep', None)
-        recursive = opts.pop('recursive', None)
-        method = opts.pop('method', 'matching')
-    else:
-        method = 'old'
-
-    def groebnersimp(ex, **opts):
-        def traverse(e):
-            if e.is_Atom:
-                return e
-            args = [traverse(x) for x in e.args]
-            if e.is_Function or e.is_Pow:
-                args = [trigsimp_groebner(x, **opts) for x in args]
-            return e.func(*args)
-        new = traverse(ex)
-        if not isinstance(new, Expr):
-            return new
-        return trigsimp_groebner(new, **opts)
-
-    trigsimpfunc = {
-        'fu': (lambda x: fu(x, **opts)),
-        'matching': (lambda x: futrig(x)),
-        'groebner': (lambda x: groebnersimp(x, **opts)),
-        'combined': (lambda x: futrig(groebnersimp(x,
-                               polynomial=True, hints=[2, tan]))),
-        'old': lambda x: trigsimp_old(x, **opts),
-                   }[method]
-
-    return trigsimpfunc(expr)
-
-
-def collect_sqrt(expr, evaluate=None):
-    """Return expr with terms having common square roots collected together.
-    If ``evaluate`` is False a count indicating the number of sqrt-containing
-    terms will be returned and, if non-zero, the terms of the Add will be
-    returned, else the expression itself will be returned as a single term.
-    If ``evaluate`` is True, the expression with any collected terms will be
-    returned.
-
-    Note: since I = sqrt(-1), it is collected, too.
-
-    Examples
-    ========
-
-    >>> from sympy import sqrt
-    >>> from sympy.simplify.simplify import collect_sqrt
-    >>> from sympy.abc import a, b
-
-    >>> r2, r3, r5 = [sqrt(i) for i in [2, 3, 5]]
-    >>> collect_sqrt(a*r2 + b*r2)
-    sqrt(2)*(a + b)
-    >>> collect_sqrt(a*r2 + b*r2 + a*r3 + b*r3)
-    sqrt(2)*(a + b) + sqrt(3)*(a + b)
-    >>> collect_sqrt(a*r2 + b*r2 + a*r3 + b*r5)
-    sqrt(3)*a + sqrt(5)*b + sqrt(2)*(a + b)
-
-    If evaluate is False then the arguments will be sorted and
-    returned as a list and a count of the number of sqrt-containing
-    terms will be returned:
-
-    >>> collect_sqrt(a*r2 + b*r2 + a*r3 + b*r5, evaluate=False)
-    ((sqrt(3)*a, sqrt(5)*b, sqrt(2)*(a + b)), 3)
-    >>> collect_sqrt(a*sqrt(2) + b, evaluate=False)
-    ((b, sqrt(2)*a), 1)
-    >>> collect_sqrt(a + b, evaluate=False)
-    ((a + b,), 0)
-
-    See Also
-    ========
-    collect, collect_const, rcollect
-    """
-    if evaluate is None:
-        evaluate = global_evaluate[0]
-    # this step will help to standardize any complex arguments
-    # of sqrts
-    coeff, expr = expr.as_content_primitive()
-    vars = set()
-    for a in Add.make_args(expr):
-        for m in a.args_cnc()[0]:
-            if m.is_number and (
-                    m.is_Pow and m.exp.is_Rational and m.exp.q == 2 or
-                    m is S.ImaginaryUnit):
-                vars.add(m)
-
-    # we only want radicals, so exclude Number handling; in this case
-    # d will be evaluated
-    d = collect_const(expr, *vars, Numbers=False)
-    hit = expr != d
-
-    if not evaluate:
-        nrad = 0
-        # make the evaluated args canonical
-        args = list(ordered(Add.make_args(d)))
-        for i, m in enumerate(args):
-            c, nc = m.args_cnc()
-            for ci in c:
-                # XXX should this be restricted to ci.is_number as above?
-                if ci.is_Pow and ci.exp.is_Rational and ci.exp.q == 2 or \
-                        ci is S.ImaginaryUnit:
-                    nrad += 1
-                    break
-            args[i] *= coeff
-        if not (hit or nrad):
-            args = [Add(*args)]
-        return tuple(args), nrad
-
-    return coeff*d
-
-
-def collect_const(expr, *vars, **kwargs):
-    """A non-greedy collection of terms with similar number coefficients in
-    an Add expr. If ``vars`` is given then only those constants will be
-    targeted. Although any Number can also be targeted, if this is not
-    desired set ``Numbers=False`` and no Float or Rational will be collected.
-
-    Examples
-    ========
-
-    >>> from sympy import sqrt
-    >>> from sympy.abc import a, s, x, y, z
-    >>> from sympy.simplify.simplify import collect_const
-    >>> collect_const(sqrt(3) + sqrt(3)*(1 + sqrt(2)))
-    sqrt(3)*(sqrt(2) + 2)
-    >>> collect_const(sqrt(3)*s + sqrt(7)*s + sqrt(3) + sqrt(7))
-    (sqrt(3) + sqrt(7))*(s + 1)
-    >>> s = sqrt(2) + 2
-    >>> collect_const(sqrt(3)*s + sqrt(3) + sqrt(7)*s + sqrt(7))
-    (sqrt(2) + 3)*(sqrt(3) + sqrt(7))
-    >>> collect_const(sqrt(3)*s + sqrt(3) + sqrt(7)*s + sqrt(7), sqrt(3))
-    sqrt(7) + sqrt(3)*(sqrt(2) + 3) + sqrt(7)*(sqrt(2) + 2)
-
-    The collection is sign-sensitive, giving higher precedence to the
-    unsigned values:
-
-    >>> collect_const(x - y - z)
-    x - (y + z)
-    >>> collect_const(-y - z)
-    -(y + z)
-    >>> collect_const(2*x - 2*y - 2*z, 2)
-    2*(x - y - z)
-    >>> collect_const(2*x - 2*y - 2*z, -2)
-    2*x - 2*(y + z)
-
-    See Also
-    ========
-    collect, collect_sqrt, rcollect
-    """
-    if not expr.is_Add:
-        return expr
-
-    recurse = False
-    Numbers = kwargs.get('Numbers', True)
-
-    if not vars:
-        recurse = True
-        vars = set()
-        for a in expr.args:
-            for m in Mul.make_args(a):
-                if m.is_number:
-                    vars.add(m)
-    else:
-        vars = sympify(vars)
-    if not Numbers:
-        vars = [v for v in vars if not v.is_Number]
-
-    vars = list(ordered(vars))
-    for v in vars:
-        terms = defaultdict(list)
-        Fv = Factors(v)
-        for m in Add.make_args(expr):
-            f = Factors(m)
-            q, r = f.div(Fv)
-            if r.is_one:
-                # only accept this as a true factor if
-                # it didn't change an exponent from an Integer
-                # to a non-Integer, e.g. 2/sqrt(2) -> sqrt(2)
-                # -- we aren't looking for this sort of change
-                fwas = f.factors.copy()
-                fnow = q.factors
-                if not any(k in fwas and fwas[k].is_Integer and not
-                        fnow[k].is_Integer for k in fnow):
-                    terms[v].append(q.as_expr())
-                    continue
-            terms[S.One].append(m)
-
-        args = []
-        hit = False
-        uneval = False
-        for k in ordered(terms):
-            v = terms[k]
-            if k is S.One:
-                args.extend(v)
-                continue
-
-            if len(v) > 1:
-                v = Add(*v)
-                hit = True
-                if recurse and v != expr:
-                    vars.append(v)
-            else:
-                v = v[0]
-
-            # be careful not to let uneval become True unless
-            # it must be because it's going to be more expensive
-            # to rebuild the expression as an unevaluated one
-            if Numbers and k.is_Number and v.is_Add:
-                args.append(_keep_coeff(k, v, sign=True))
-                uneval = True
-            else:
-                args.append(k*v)
-
-        if hit:
-            if uneval:
-                expr = _unevaluated_Add(*args)
-            else:
-                expr = Add(*args)
-            if not expr.is_Add:
-                break
-
-    return expr
-
-
-def _split_gcd(*a):
-    """
-    split the list of integers ``a`` into a list of integers, ``a1`` having
-    ``g = gcd(a1)``, and a list ``a2`` whose elements are not divisible by
-    ``g``.  Returns ``g, a1, a2``
-
-    Examples
-    ========
-
-    >>> from sympy.simplify.simplify import _split_gcd
-    >>> _split_gcd(55, 35, 22, 14, 77, 10)
-    (5, [55, 35, 10], [22, 14, 77])
-    """
-    g = a[0]
-    b1 = [g]
-    b2 = []
-    for x in a[1:]:
-        g1 = gcd(g, x)
-        if g1 == 1:
-            b2.append(x)
-        else:
-            g = g1
-            b1.append(x)
-    return g, b1, b2
-
-def _is_sum_surds(p):
-    args = p.args if p.is_Add else [p]
-    for y in args:
-        if not ((y**2).is_Rational and y.is_real):
-            return False
-    return True
-
-def _nthroot_solve(p, n, prec):
-    """
-     helper function for ``nthroot``
-     It denests ``p**Rational(1, n)`` using its minimal polynomial
-    """
-    from sympy.polys.numberfields import _minimal_polynomial_sq
-    from sympy.solvers import solve
-    while n % 2 == 0:
-        p = sqrtdenest(sqrt(p))
-        n = n // 2
-    if n == 1:
-        return p
-    pn = p**Rational(1, n)
-    x = Symbol('x')
-    f = _minimal_polynomial_sq(p, n, x)
-    if f is None:
-        return None
-    sols = solve(f, x)
-    for sol in sols:
-        if abs(sol - pn).n() < 1./10**prec:
-            sol = sqrtdenest(sol)
-            if _mexpand(sol**n) == p:
-                return sol
-
-def nthroot(expr, n, max_len=4, prec=15):
-    """
-    compute a real nth-root of a sum of surds
-
-    Parameters
-    ==========
-
-    expr : sum of surds
-    n : integer
-    max_len : maximum number of surds passed as constants to ``nsimplify``
-
-    Algorithm
-    =========
-
-    First ``nsimplify`` is used to get a candidate root; if it is not a
-    root the minimal polynomial is computed; the answer is one of its
-    roots.
-
-    Examples
-    ========
-
-    >>> from sympy.simplify.simplify import nthroot
-    >>> from sympy import Rational, sqrt
-    >>> nthroot(90 + 34*sqrt(7), 3)
-    sqrt(7) + 3
-
-    """
-    expr = sympify(expr)
-    n = sympify(n)
-    p = expr**Rational(1, n)
-    if not n.is_integer:
-        return p
-    if not _is_sum_surds(expr):
-        return p
-    surds = []
-    coeff_muls = [x.as_coeff_Mul() for x in expr.args]
-    for x, y in coeff_muls:
-        if not x.is_rational:
-            return p
-        if y is S.One:
-            continue
-        if not (y.is_Pow and y.exp == S.Half and y.base.is_integer):
-            return p
-        surds.append(y)
-    surds.sort()
-    surds = surds[:max_len]
-    if expr < 0 and n % 2 == 1:
-        p = (-expr)**Rational(1, n)
-        a = nsimplify(p, constants=surds)
-        res = a if _mexpand(a**n) == _mexpand(-expr) else p
-        return -res
-    a = nsimplify(p, constants=surds)
-    if _mexpand(a) is not _mexpand(p) and _mexpand(a**n) == _mexpand(expr):
-        return _mexpand(a)
-    expr = _nthroot_solve(expr, n, prec)
-    if expr is None:
-        return p
-    return expr
-
-
-def split_surds(expr):
-    """
-    split an expression with terms whose squares are rationals
-    into a sum of terms whose surds squared have gcd equal to g
-    and a sum of terms with surds squared prime with g
-
-    Examples
-    ========
-
-    >>> from sympy import sqrt
-    >>> from sympy.simplify.simplify import split_surds
-    >>> split_surds(3*sqrt(3) + sqrt(5)/7 + sqrt(6) + sqrt(10) + sqrt(15))
-    (3, sqrt(2) + sqrt(5) + 3, sqrt(5)/7 + sqrt(10))
-    """
-    args = sorted(expr.args, key=default_sort_key)
-    coeff_muls = [x.as_coeff_Mul() for x in args]
-    surds = [x[1]**2 for x in coeff_muls if x[1].is_Pow]
-    surds.sort(key=default_sort_key)
-    g, b1, b2 = _split_gcd(*surds)
-    g2 = g
-    if not b2 and len(b1) >= 2:
-        b1n = [x/g for x in b1]
-        b1n = [x for x in b1n if x != 1]
-        # only a common factor has been factored; split again
-        g1, b1n, b2 = _split_gcd(*b1n)
-        g2 = g*g1
-    a1v, a2v = [], []
-    for c, s in coeff_muls:
-        if s.is_Pow and s.exp == S.Half:
-            s1 = s.base
-            if s1 in b1:
-                a1v.append(c*sqrt(s1/g2))
-            else:
-                a2v.append(c*s)
-        else:
-            a2v.append(c*s)
-    a = Add(*a1v)
-    b = Add(*a2v)
-    return g2, a, b
-
-
-def rad_rationalize(num, den):
-    """
-    Rationalize num/den by removing square roots in the denominator;
-    num and den are sum of terms whose squares are rationals
-
-    Examples
-    ========
-
-    >>> from sympy import sqrt
-    >>> from sympy.simplify.simplify import rad_rationalize
-    >>> rad_rationalize(sqrt(3), 1 + sqrt(2)/3)
-    (-sqrt(3) + sqrt(6)/3, -7/9)
-    """
-    if not den.is_Add:
-        return num, den
-    g, a, b = split_surds(den)
-    a = a*sqrt(g)
-    num = _mexpand((a - b)*num)
-    den = _mexpand(a**2 - b**2)
-    return rad_rationalize(num, den)
-
-
-def radsimp(expr, symbolic=True, max_terms=4):
-    """
-    Rationalize the denominator by removing square roots.
-
-    Note: the expression returned from radsimp must be used with caution
-    since if the denominator contains symbols, it will be possible to make
-    substitutions that violate the assumptions of the simplification process:
-    that for a denominator matching a + b*sqrt(c), a != +/-b*sqrt(c). (If
-    there are no symbols, this assumptions is made valid by collecting terms
-    of sqrt(c) so the match variable ``a`` does not contain ``sqrt(c)``.) If
-    you do not want the simplification to occur for symbolic denominators, set
-    ``symbolic`` to False.
-
-    If there are more than ``max_terms`` radical terms then the expression is
-    returned unchanged.
-
-    Examples
-    ========
-
-    >>> from sympy import radsimp, sqrt, Symbol, denom, pprint, I
-    >>> from sympy import factor_terms, fraction, signsimp
-    >>> from sympy.simplify.simplify import collect_sqrt
-    >>> from sympy.abc import a, b, c
-
-    >>> radsimp(1/(I + 1))
-    (1 - I)/2
-    >>> radsimp(1/(2 + sqrt(2)))
-    (-sqrt(2) + 2)/2
-    >>> x,y = map(Symbol, 'xy')
-    >>> e = ((2 + 2*sqrt(2))*x + (2 + sqrt(8))*y)/(2 + sqrt(2))
-    >>> radsimp(e)
-    sqrt(2)*(x + y)
-
-    No simplification beyond removal of the gcd is done. One might
-    want to polish the result a little, however, by collecting
-    square root terms:
-
-    >>> r2 = sqrt(2)
-    >>> r5 = sqrt(5)
-    >>> ans = radsimp(1/(y*r2 + x*r2 + a*r5 + b*r5)); pprint(ans)
-        ___       ___       ___       ___
-      \/ 5 *a + \/ 5 *b - \/ 2 *x - \/ 2 *y
-    ------------------------------------------
-       2               2      2              2
-    5*a  + 10*a*b + 5*b  - 2*x  - 4*x*y - 2*y
-
-    >>> n, d = fraction(ans)
-    >>> pprint(factor_terms(signsimp(collect_sqrt(n))/d, radical=True))
-            ___             ___
-          \/ 5 *(a + b) - \/ 2 *(x + y)
-    ------------------------------------------
-       2               2      2              2
-    5*a  + 10*a*b + 5*b  - 2*x  - 4*x*y - 2*y
-
-    If radicals in the denominator cannot be removed or there is no denominator,
-    the original expression will be returned.
-
-    >>> radsimp(sqrt(2)*x + sqrt(2))
-    sqrt(2)*x + sqrt(2)
-
-    Results with symbols will not always be valid for all substitutions:
-
-    >>> eq = 1/(a + b*sqrt(c))
-    >>> eq.subs(a, b*sqrt(c))
-    1/(2*b*sqrt(c))
-    >>> radsimp(eq).subs(a, b*sqrt(c))
-    nan
-
-    If symbolic=False, symbolic denominators will not be transformed (but
-    numeric denominators will still be processed):
-
-    >>> radsimp(eq, symbolic=False)
-    1/(a + b*sqrt(c))
-
-    """
-
-    syms = symbols("a:d A:D")
-    def _num(rterms):
-        # return the multiplier that will simplify the expression described
-        # by rterms [(sqrt arg, coeff), ... ]
-        a, b, c, d, A, B, C, D = syms
-        if len(rterms) == 2:
-            reps = dict(list(zip([A, a, B, b], [j for i in rterms for j in i])))
-            return (
-            sqrt(A)*a - sqrt(B)*b).xreplace(reps)
-        if len(rterms) == 3:
-            reps = dict(list(zip([A, a, B, b, C, c], [j for i in rterms for j in i])))
-            return (
-            (sqrt(A)*a + sqrt(B)*b - sqrt(C)*c)*(2*sqrt(A)*sqrt(B)*a*b - A*a**2 -
-            B*b**2 + C*c**2)).xreplace(reps)
-        elif len(rterms) == 4:
-            reps = dict(list(zip([A, a, B, b, C, c, D, d], [j for i in rterms for j in i])))
-            return ((sqrt(A)*a + sqrt(B)*b - sqrt(C)*c - sqrt(D)*d)*(2*sqrt(A)*sqrt(B)*a*b
-                - A*a**2 - B*b**2 - 2*sqrt(C)*sqrt(D)*c*d + C*c**2 +
-                D*d**2)*(-8*sqrt(A)*sqrt(B)*sqrt(C)*sqrt(D)*a*b*c*d + A**2*a**4 -
-                2*A*B*a**2*b**2 - 2*A*C*a**2*c**2 - 2*A*D*a**2*d**2 + B**2*b**4 -
-                2*B*C*b**2*c**2 - 2*B*D*b**2*d**2 + C**2*c**4 - 2*C*D*c**2*d**2 +
-                D**2*d**4)).xreplace(reps)
-        elif len(rterms) == 1:
-            return sqrt(rterms[0][0])
-        else:
-            raise NotImplementedError
-
-    def ispow2(d, log2=False):
-        if not d.is_Pow:
-            return False
-        e = d.exp
-        if e.is_Rational and e.q == 2 or symbolic and fraction(e)[1] == 2:
-            return True
-        if log2:
-            q = 1
-            if e.is_Rational:
-                q = e.q
-            elif symbolic:
-                d = fraction(e)[1]
-                if d.is_Integer:
-                    q = d
-            if q != 1 and log(q, 2).is_Integer:
-                return True
-        return False
-
-    def handle(expr):
-        # Handle first reduces to the case
-        # expr = 1/d, where d is an add, or d is base**p/2.
-        # We do this by recursively calling handle on each piece.
-        n, d = fraction(expr)
-
-        if expr.is_Atom or (d.is_Atom and n.is_Atom):
-            return expr
-        elif not n.is_Atom:
-            n = n.func(*[handle(a) for a in n.args])
-            return _unevaluated_Mul(n, handle(1/d))
-        elif n is not S.One:
-            return _unevaluated_Mul(n, handle(1/d))
-        elif d.is_Mul:
-            return _unevaluated_Mul(*[handle(1/d) for d in d.args])
-
-        # By this step, expr is 1/d, and d is not a mul.
-        if not symbolic and d.free_symbols:
-            return expr
-
-        if ispow2(d):
-            d2 = sqrtdenest(sqrt(d.base))**fraction(d.exp)[0]
-            if d2 != d:
-                return handle(1/d2)
-        elif d.is_Pow and (d.exp.is_integer or d.base.is_positive):
-            # (1/d**i) = (1/d)**i
-            return handle(1/d.base)**d.exp
-
-        if not (d.is_Add or ispow2(d)):
-            return 1/d.func(*[handle(a) for a in d.args])
-
-        # handle 1/d treating d as an Add (though it may not be)
-
-        keep = True  # keep changes that are made
-
-        # flatten it and collect radicals after checking for special
-        # conditions
-        d = _mexpand(d)
-
-        # did it change?
-        if d.is_Atom:
-            return 1/d
-
-        # is it a number that might be handled easily?
-        if d.is_number:
-            _d = nsimplify(d)
-            if _d.is_Number and _d.equals(d):
-                return 1/_d
-
-        while True:
-            # collect similar terms
-            collected = defaultdict(list)
-            for m in Add.make_args(d):  # d might have become non-Add
-                p2 = []
-                other = []
-                for i in Mul.make_args(m):
-                    if ispow2(i, log2=True):
-                        p2.append(i.base if i.exp is S.Half else i.base**(2*i.exp))
-                    elif i is S.ImaginaryUnit:
-                        p2.append(S.NegativeOne)
-                    else:
-                        other.append(i)
-                collected[tuple(ordered(p2))].append(Mul(*other))
-            rterms = list(ordered(list(collected.items())))
-            rterms = [(Mul(*i), Add(*j)) for i, j in rterms]
-            nrad = len(rterms) - (1 if rterms[0][0] is S.One else 0)
-            if nrad < 1:
-                break
-            elif nrad > max_terms:
-                # there may have been invalid operations leading to this point
-                # so don't keep changes, e.g. this expression is troublesome
-                # in collecting terms so as not to raise the issue of 2834:
-                # r = sqrt(sqrt(5) + 5)
-                # eq = 1/(sqrt(5)*r + 2*sqrt(5)*sqrt(-sqrt(5) + 5) + 5*r)
-                keep = False
-                break
-            if len(rterms) > 4:
-                # in general, only 4 terms can be removed with repeated squaring
-                # but other considerations can guide selection of radical terms
-                # so that radicals are removed
-                if all([x.is_Integer and (y**2).is_Rational for x, y in rterms]):
-                    nd, d = rad_rationalize(S.One, Add._from_args(
-                        [sqrt(x)*y for x, y in rterms]))
-                    n *= nd
-                else:
-                    # is there anything else that might be attempted?
-                    keep = False
-                break
-
-            num = powsimp(_num(rterms))
-            n *= num
-            d *= num
-            d = powdenest(_mexpand(d), force=symbolic)
-            if d.is_Atom:
-                break
-
-        if not keep:
-            return expr
-        return _unevaluated_Mul(n, 1/d)
-
-    coeff, expr = expr.as_coeff_Add()
-    expr = expr.normal()
-    old = fraction(expr)
-    n, d = fraction(handle(expr))
-    if old != (n, d):
-        if not d.is_Atom:
-            was = (n, d)
-            n = signsimp(n, evaluate=False)
-            d = signsimp(d, evaluate=False)
-            u = Factors(_unevaluated_Mul(n, 1/d))
-            u = _unevaluated_Mul(*[k**v for k, v in u.factors.items()])
-            n, d = fraction(u)
-            if old == (n, d):
-                n, d = was
-        n = expand_mul(n)
-        if d.is_Number or d.is_Add:
-            n2, d2 = fraction(gcd_terms(_unevaluated_Mul(n, 1/d)))
-            if d2.is_Number or (d2.count_ops() <= d.count_ops()):
-                n, d = [signsimp(i) for i in (n2, d2)]
-                if n.is_Mul and n.args[0].is_Number:
-                    n = n.func(*n.args)
-
-    return coeff + _unevaluated_Mul(n, 1/d)
-
-
-def posify(eq):
-    """Return eq (with generic symbols made positive) and a restore
-    dictionary.
-
-    Any symbol that has positive=None will be replaced with a positive dummy
-    symbol having the same name. This replacement will allow more symbolic
-    processing of expressions, especially those involving powers and
-    logarithms.
-
-    A dictionary that can be sent to subs to restore eq to its original
-    symbols is also returned.
-
-    >>> from sympy import posify, Symbol, log
-    >>> from sympy.abc import x
-    >>> posify(x + Symbol('p', positive=True) + Symbol('n', negative=True))
-    (_x + n + p, {_x: x})
-
-    >> log(1/x).expand() # should be log(1/x) but it comes back as -log(x)
-    log(1/x)
-
-    >>> log(posify(1/x)[0]).expand() # take [0] and ignore replacements
-    -log(_x)
-    >>> eq, rep = posify(1/x)
-    >>> log(eq).expand().subs(rep)
-    -log(x)
-    >>> posify([x, 1 + x])
-    ([_x, _x + 1], {_x: x})
-    """
-    eq = sympify(eq)
-    if iterable(eq):
-        f = type(eq)
-        eq = list(eq)
-        syms = set()
-        for e in eq:
-            syms = syms.union(e.atoms(Symbol))
-        reps = {}
-        for s in syms:
-            reps.update(dict((v, k) for k, v in posify(s)[1].items()))
-        for i, e in enumerate(eq):
-            eq[i] = e.subs(reps)
-        return f(eq), dict([(r, s) for s, r in reps.items()])
-
-    reps = dict([(s, Dummy(s.name, positive=True))
-                 for s in eq.free_symbols if s.is_positive is None])
-    eq = eq.subs(reps)
-    return eq, dict([(r, s) for s, r in reps.items()])
-
-
 def _polarify(eq, lift, pause=False):
     from sympy import Integral
     if eq.is_polar:
@@ -2239,686 +562,60 @@ def unpolarify(eq, subs={}, exponents_only=False):
     return res.subs({exp_polar(0): 1, polar_lift(0): 0})
 
 
-def _denest_pow(eq):
+def _is_sum_surds(p):
+    args = p.args if p.is_Add else [p]
+    for y in args:
+        if not ((y**2).is_Rational and y.is_real):
+            return False
+    return True
+
+
+def posify(eq):
+    """Return eq (with generic symbols made positive) and a restore
+    dictionary.
+
+    Any symbol that has positive=None will be replaced with a positive dummy
+    symbol having the same name. This replacement will allow more symbolic
+    processing of expressions, especially those involving powers and
+    logarithms.
+
+    A dictionary that can be sent to subs to restore eq to its original
+    symbols is also returned.
+
+    >>> from sympy import posify, Symbol, log
+    >>> from sympy.abc import x
+    >>> posify(x + Symbol('p', positive=True) + Symbol('n', negative=True))
+    (_x + n + p, {_x: x})
+
+    >> log(1/x).expand() # should be log(1/x) but it comes back as -log(x)
+    log(1/x)
+
+    >>> log(posify(1/x)[0]).expand() # take [0] and ignore replacements
+    -log(_x)
+    >>> eq, rep = posify(1/x)
+    >>> log(eq).expand().subs(rep)
+    -log(x)
+    >>> posify([x, 1 + x])
+    ([_x, _x + 1], {_x: x})
     """
-    Denest powers.
-
-    This is a helper function for powdenest that performs the actual
-    transformation.
-    """
-    b, e = eq.as_base_exp()
-    if b.is_Pow or isinstance(b.func, exp) and e != 1:
-        new = b._eval_power(e)
-        if new is not None:
-            eq = new
-            b, e = new.as_base_exp()
-
-    # denest exp with log terms in exponent
-    if b is S.Exp1 and e.is_Mul:
-        logs = []
-        other = []
-        for ei in e.args:
-            if any(ai.func is log for ai in Add.make_args(ei)):
-                logs.append(ei)
-            else:
-                other.append(ei)
-        logs = logcombine(Mul(*logs))
-        return Pow(exp(logs), Mul(*other))
-
-    _, be = b.as_base_exp()
-    if be is S.One and not (b.is_Mul or
-                            b.is_Rational and b.q != 1 or
-                            b.is_positive):
-        return eq
-
-    # denest eq which is either pos**e or Pow**e or Mul**e or
-    # Mul(b1**e1, b2**e2)
-
-    # handle polar numbers specially
-    polars, nonpolars = [], []
-    for bb in Mul.make_args(b):
-        if bb.is_polar:
-            polars.append(bb.as_base_exp())
-        else:
-            nonpolars.append(bb)
-    if len(polars) == 1 and not polars[0][0].is_Mul:
-        return Pow(polars[0][0], polars[0][1]*e)*powdenest(Mul(*nonpolars)**e)
-    elif polars:
-        return Mul(*[powdenest(bb**(ee*e)) for (bb, ee) in polars]) \
-            *powdenest(Mul(*nonpolars)**e)
-
-    if b.is_Integer:
-        # use log to see if there is a power here
-        logb = expand_log(log(b))
-        if logb.is_Mul:
-            c, logb = logb.args
-            e *= c
-            base = logb.args[0]
-            return Pow(base, e)
-
-    # if b is not a Mul or any factor is an atom then there is nothing to do
-    if not b.is_Mul or any(s.is_Atom for s in Mul.make_args(b)):
-        return eq
-
-    # let log handle the case of the base of the argument being a Mul, e.g.
-    # sqrt(x**(2*i)*y**(6*i)) -> x**i*y**(3**i) if x and y are positive; we
-    # will take the log, expand it, and then factor out the common powers that
-    # now appear as coefficient. We do this manually since terms_gcd pulls out
-    # fractions, terms_gcd(x+x*y/2) -> x*(y + 2)/2 and we don't want the 1/2;
-    # gcd won't pull out numerators from a fraction: gcd(3*x, 9*x/2) -> x but
-    # we want 3*x. Neither work with noncommutatives.
-
-    def nc_gcd(aa, bb):
-        a, b = [i.as_coeff_Mul() for i in [aa, bb]]
-        c = gcd(a[0], b[0]).as_numer_denom()[0]
-        g = Mul(*(a[1].args_cnc(cset=True)[0] & b[1].args_cnc(cset=True)[0]))
-        return _keep_coeff(c, g)
-
-    glogb = expand_log(log(b))
-    if glogb.is_Add:
-        args = glogb.args
-        g = reduce(nc_gcd, args)
-        if g != 1:
-            cg, rg = g.as_coeff_Mul()
-            glogb = _keep_coeff(cg, rg*Add(*[a/g for a in args]))
-
-    # now put the log back together again
-    if glogb.func is log or not glogb.is_Mul:
-        if glogb.args[0].is_Pow or glogb.args[0].func is exp:
-            glogb = _denest_pow(glogb.args[0])
-            if (abs(glogb.exp) < 1) == True:
-                return Pow(glogb.base, glogb.exp*e)
-        return eq
-
-    # the log(b) was a Mul so join any adds with logcombine
-    add = []
-    other = []
-    for a in glogb.args:
-        if a.is_Add:
-            add.append(a)
-        else:
-            other.append(a)
-    return Pow(exp(logcombine(Mul(*add))), e*Mul(*other))
-
-
-def powdenest(eq, force=False, polar=False):
-    r"""
-    Collect exponents on powers as assumptions allow.
-
-    Given ``(bb**be)**e``, this can be simplified as follows:
-        * if ``bb`` is positive, or
-        * ``e`` is an integer, or
-        * ``|be| < 1`` then this simplifies to ``bb**(be*e)``
-
-    Given a product of powers raised to a power, ``(bb1**be1 *
-    bb2**be2...)**e``, simplification can be done as follows:
-
-    - if e is positive, the gcd of all bei can be joined with e;
-    - all non-negative bb can be separated from those that are negative
-      and their gcd can be joined with e; autosimplification already
-      handles this separation.
-    - integer factors from powers that have integers in the denominator
-      of the exponent can be removed from any term and the gcd of such
-      integers can be joined with e
-
-    Setting ``force`` to True will make symbols that are not explicitly
-    negative behave as though they are positive, resulting in more
-    denesting.
-
-    Setting ``polar`` to True will do simplifications on the Riemann surface of
-    the logarithm, also resulting in more denestings.
-
-    When there are sums of logs in exp() then a product of powers may be
-    obtained e.g. ``exp(3*(log(a) + 2*log(b)))`` - > ``a**3*b**6``.
-
-    Examples
-    ========
-
-    >>> from sympy.abc import a, b, x, y, z
-    >>> from sympy import Symbol, exp, log, sqrt, symbols, powdenest
-
-    >>> powdenest((x**(2*a/3))**(3*x))
-    (x**(2*a/3))**(3*x)
-    >>> powdenest(exp(3*x*log(2)))
-    2**(3*x)
-
-    Assumptions may prevent expansion:
-
-    >>> powdenest(sqrt(x**2))
-    sqrt(x**2)
-
-    >>> p = symbols('p', positive=True)
-    >>> powdenest(sqrt(p**2))
-    p
-
-    No other expansion is done.
-
-    >>> i, j = symbols('i,j', integer=True)
-    >>> powdenest((x**x)**(i + j)) # -X-> (x**x)**i*(x**x)**j
-    x**(x*(i + j))
-
-    But exp() will be denested by moving all non-log terms outside of
-    the function; this may result in the collapsing of the exp to a power
-    with a different base:
-
-    >>> powdenest(exp(3*y*log(x)))
-    x**(3*y)
-    >>> powdenest(exp(y*(log(a) + log(b))))
-    (a*b)**y
-    >>> powdenest(exp(3*(log(a) + log(b))))
-    a**3*b**3
-
-    If assumptions allow, symbols can also be moved to the outermost exponent:
-
-    >>> i = Symbol('i', integer=True)
-    >>> powdenest(((x**(2*i))**(3*y))**x)
-    ((x**(2*i))**(3*y))**x
-    >>> powdenest(((x**(2*i))**(3*y))**x, force=True)
-    x**(6*i*x*y)
-
-    >>> powdenest(((x**(2*a/3))**(3*y/i))**x)
-    ((x**(2*a/3))**(3*y/i))**x
-    >>> powdenest((x**(2*i)*y**(4*i))**z, force=True)
-    (x*y**2)**(2*i*z)
-
-    >>> n = Symbol('n', negative=True)
-
-    >>> powdenest((x**i)**y, force=True)
-    x**(i*y)
-    >>> powdenest((n**i)**x, force=True)
-    (n**i)**x
-
-    """
-
-    if force:
-        eq, rep = posify(eq)
-        return powdenest(eq, force=False).xreplace(rep)
-
-    if polar:
-        eq, rep = polarify(eq)
-        return unpolarify(powdenest(unpolarify(eq, exponents_only=True)), rep)
-
-    new = powsimp(sympify(eq))
-    return new.xreplace(Transform(
-        _denest_pow, filter=lambda m: m.is_Pow or m.func is exp))
-
-_y = Dummy('y')
-
-
-def powsimp(expr, deep=False, combine='all', force=False, measure=count_ops):
-    """
-    reduces expression by combining powers with similar bases and exponents.
-
-    Notes
-    =====
-
-    If deep is True then powsimp() will also simplify arguments of
-    functions. By default deep is set to False.
-
-    If force is True then bases will be combined without checking for
-    assumptions, e.g. sqrt(x)*sqrt(y) -> sqrt(x*y) which is not true
-    if x and y are both negative.
-
-    You can make powsimp() only combine bases or only combine exponents by
-    changing combine='base' or combine='exp'.  By default, combine='all',
-    which does both.  combine='base' will only combine::
-
-         a   a          a                          2x      x
-        x * y  =>  (x*y)   as well as things like 2   =>  4
-
-    and combine='exp' will only combine
-    ::
-
-         a   b      (a + b)
-        x * x  =>  x
-
-    combine='exp' will strictly only combine exponents in the way that used
-    to be automatic.  Also use deep=True if you need the old behavior.
-
-    When combine='all', 'exp' is evaluated first.  Consider the first
-    example below for when there could be an ambiguity relating to this.
-    This is done so things like the second example can be completely
-    combined.  If you want 'base' combined first, do something like
-    powsimp(powsimp(expr, combine='base'), combine='exp').
-
-    Examples
-    ========
-
-    >>> from sympy import powsimp, exp, log, symbols
-    >>> from sympy.abc import x, y, z, n
-    >>> powsimp(x**y*x**z*y**z, combine='all')
-    x**(y + z)*y**z
-    >>> powsimp(x**y*x**z*y**z, combine='exp')
-    x**(y + z)*y**z
-    >>> powsimp(x**y*x**z*y**z, combine='base', force=True)
-    x**y*(x*y)**z
-
-    >>> powsimp(x**z*x**y*n**z*n**y, combine='all', force=True)
-    (n*x)**(y + z)
-    >>> powsimp(x**z*x**y*n**z*n**y, combine='exp')
-    n**(y + z)*x**(y + z)
-    >>> powsimp(x**z*x**y*n**z*n**y, combine='base', force=True)
-    (n*x)**y*(n*x)**z
-
-    >>> x, y = symbols('x y', positive=True)
-    >>> powsimp(log(exp(x)*exp(y)))
-    log(exp(x)*exp(y))
-    >>> powsimp(log(exp(x)*exp(y)), deep=True)
-    x + y
-
-    Radicals with Mul bases will be combined if combine='exp'
-
-    >>> from sympy import sqrt, Mul
-    >>> x, y = symbols('x y')
-
-    Two radicals are automatically joined through Mul:
-
-    >>> a=sqrt(x*sqrt(y))
-    >>> a*a**3 == a**4
-    True
-
-    But if an integer power of that radical has been
-    autoexpanded then Mul does not join the resulting factors:
-
-    >>> a**4 # auto expands to a Mul, no longer a Pow
-    x**2*y
-    >>> _*a # so Mul doesn't combine them
-    x**2*y*sqrt(x*sqrt(y))
-    >>> powsimp(_) # but powsimp will
-    (x*sqrt(y))**(5/2)
-    >>> powsimp(x*y*a) # but won't when doing so would violate assumptions
-    x*y*sqrt(x*sqrt(y))
-
-    """
-    from sympy.matrices.expressions.matexpr import MatrixSymbol
-
-    def recurse(arg, **kwargs):
-        _deep = kwargs.get('deep', deep)
-        _combine = kwargs.get('combine', combine)
-        _force = kwargs.get('force', force)
-        _measure = kwargs.get('measure', measure)
-        return powsimp(arg, _deep, _combine, _force, _measure)
-
-    expr = sympify(expr)
-
-    if (not isinstance(expr, Basic) or isinstance(expr, MatrixSymbol) or (
-            expr.is_Atom or expr in (exp_polar(0), exp_polar(1)))):
-        return expr
-
-    if deep or expr.is_Add or expr.is_Mul and _y not in expr.args:
-        expr = expr.func(*[recurse(w) for w in expr.args])
-
-    if expr.is_Pow:
-        return recurse(expr*_y, deep=False)/_y
-
-    if not expr.is_Mul:
-        return expr
-
-    # handle the Mul
-    if combine in ('exp', 'all'):
-        # Collect base/exp data, while maintaining order in the
-        # non-commutative parts of the product
-        c_powers = defaultdict(list)
-        nc_part = []
-        newexpr = []
-        coeff = S.One
-        for term in expr.args:
-            if term.is_Rational:
-                coeff *= term
-                continue
-            if term.is_Pow:
-                term = _denest_pow(term)
-            if term.is_commutative:
-                b, e = term.as_base_exp()
-                if deep:
-                    b, e = [recurse(i) for i in [b, e]]
-                if b.is_Pow or b.func is exp:
-                    # don't let smthg like sqrt(x**a) split into x**a, 1/2
-                    # or else it will be joined as x**(a/2) later
-                    b, e = b**e, S.One
-                c_powers[b].append(e)
-            else:
-                # This is the logic that combines exponents for equal,
-                # but non-commutative bases: A**x*A**y == A**(x+y).
-                if nc_part:
-                    b1, e1 = nc_part[-1].as_base_exp()
-                    b2, e2 = term.as_base_exp()
-                    if (b1 == b2 and
-                            e1.is_commutative and e2.is_commutative):
-                        nc_part[-1] = Pow(b1, Add(e1, e2))
-                        continue
-                nc_part.append(term)
-
-        # add up exponents of common bases
-        for b, e in ordered(iter(c_powers.items())):
-            # allow 2**x/4 -> 2**(x - 2); don't do this when b and e are
-            # Numbers since autoevaluation will undo it, e.g.
-            # 2**(1/3)/4 -> 2**(1/3 - 2) -> 2**(1/3)/4
-            if (b and b.is_Number and not all(ei.is_Number for ei in e) and \
-                    coeff is not S.One and
-                    b not in (S.One, S.NegativeOne)):
-                m = multiplicity(abs(b), abs(coeff))
-                if m:
-                    e.append(m)
-                    coeff /= b**m
-            c_powers[b] = Add(*e)
-        if coeff is not S.One:
-            if coeff in c_powers:
-                c_powers[coeff] += S.One
-            else:
-                c_powers[coeff] = S.One
-
-        # convert to plain dictionary
-        c_powers = dict(c_powers)
-
-        # check for base and inverted base pairs
-        be = list(c_powers.items())
-        skip = set()  # skip if we already saw them
-        for b, e in be:
-            if b in skip:
-                continue
-            bpos = b.is_positive or b.is_polar
-            if bpos:
-                binv = 1/b
-                if b != binv and binv in c_powers:
-                    if b.as_numer_denom()[0] is S.One:
-                        c_powers.pop(b)
-                        c_powers[binv] -= e
-                    else:
-                        skip.add(binv)
-                        e = c_powers.pop(binv)
-                        c_powers[b] -= e
-
-        # check for base and negated base pairs
-        be = list(c_powers.items())
-        _n = S.NegativeOne
-        for i, (b, e) in enumerate(be):
-            if ((-b).is_Symbol or b.is_Add) and -b in c_powers:
-                if (b.is_positive in (0, 1) or e.is_integer):
-                    c_powers[-b] += c_powers.pop(b)
-                    if _n in c_powers:
-                        c_powers[_n] += e
-                    else:
-                        c_powers[_n] = e
-
-        # filter c_powers and convert to a list
-        c_powers = [(b, e) for b, e in c_powers.items() if e]
-
-        # ==============================================================
-        # check for Mul bases of Rational powers that can be combined with
-        # separated bases, e.g. x*sqrt(x*y)*sqrt(x*sqrt(x*y)) ->
-        # (x*sqrt(x*y))**(3/2)
-        # ---------------- helper functions
-
-        def ratq(x):
-            '''Return Rational part of x's exponent as it appears in the bkey.
-            '''
-            return bkey(x)[0][1]
-
-        def bkey(b, e=None):
-            '''Return (b**s, c.q), c.p where e -> c*s. If e is not given then
-            it will be taken by using as_base_exp() on the input b.
-            e.g.
-                x**3/2 -> (x, 2), 3
-                x**y -> (x**y, 1), 1
-                x**(2*y/3) -> (x**y, 3), 2
-                exp(x/2) -> (exp(a), 2), 1
-
-            '''
-            if e is not None:  # coming from c_powers or from below
-                if e.is_Integer:
-                    return (b, S.One), e
-                elif e.is_Rational:
-                    return (b, Integer(e.q)), Integer(e.p)
-                else:
-                    c, m = e.as_coeff_Mul(rational=True)
-                    if c is not S.One:
-                        return (b**m, Integer(c.q)), Integer(c.p)
-                    else:
-                        return (b**e, S.One), S.One
-            else:
-                return bkey(*b.as_base_exp())
-
-        def update(b):
-            '''Decide what to do with base, b. If its exponent is now an
-            integer multiple of the Rational denominator, then remove it
-            and put the factors of its base in the common_b dictionary or
-            update the existing bases if necessary. If it has been zeroed
-            out, simply remove the base.
-            '''
-            newe, r = divmod(common_b[b], b[1])
-            if not r:
-                common_b.pop(b)
-                if newe:
-                    for m in Mul.make_args(b[0]**newe):
-                        b, e = bkey(m)
-                        if b not in common_b:
-                            common_b[b] = 0
-                        common_b[b] += e
-                        if b[1] != 1:
-                            bases.append(b)
-        # ---------------- end of helper functions
-
-        # assemble a dictionary of the factors having a Rational power
-        common_b = {}
-        done = []
-        bases = []
-        for b, e in c_powers:
-            b, e = bkey(b, e)
-            if b in common_b.keys():
-                common_b[b] = common_b[b] + e
-            else:
-                common_b[b] = e
-            if b[1] != 1 and b[0].is_Mul:
-                bases.append(b)
-        c_powers = [(b, e) for b, e in common_b.items() if e]
-        bases.sort(key=default_sort_key)  # this makes tie-breaking canonical
-        bases.sort(key=measure, reverse=True)  # handle longest first
-        for base in bases:
-            if base not in common_b:  # it may have been removed already
-                continue
-            b, exponent = base
-            last = False  # True when no factor of base is a radical
-            qlcm = 1  # the lcm of the radical denominators
-            while True:
-                bstart = b
-                qstart = qlcm
-
-                bb = []  # list of factors
-                ee = []  # (factor's expo. and it's current value in common_b)
-                for bi in Mul.make_args(b):
-                    bib, bie = bkey(bi)
-                    if bib not in common_b or common_b[bib] < bie:
-                        ee = bb = []  # failed
-                        break
-                    ee.append([bie, common_b[bib]])
-                    bb.append(bib)
-                if ee:
-                    # find the number of extractions possible
-                    # e.g. [(1, 2), (2, 2)] -> min(2/1, 2/2) -> 1
-                    min1 = ee[0][1]/ee[0][0]
-                    for i in range(len(ee)):
-                        rat = ee[i][1]/ee[i][0]
-                        if rat < 1:
-                            break
-                        min1 = min(min1, rat)
-                    else:
-                        # update base factor counts
-                        # e.g. if ee = [(2, 5), (3, 6)] then min1 = 2
-                        # and the new base counts will be 5-2*2 and 6-2*3
-                        for i in range(len(bb)):
-                            common_b[bb[i]] -= min1*ee[i][0]
-                            update(bb[i])
-                        # update the count of the base
-                        # e.g. x**2*y*sqrt(x*sqrt(y)) the count of x*sqrt(y)
-                        # will increase by 4 to give bkey (x*sqrt(y), 2, 5)
-                        common_b[base] += min1*qstart*exponent
-                if (last  # no more radicals in base
-                    or len(common_b) == 1  # nothing left to join with
-                    or all(k[1] == 1 for k in common_b)  # no rad's in common_b
-                        ):
-                    break
-                # see what we can exponentiate base by to remove any radicals
-                # so we know what to search for
-                # e.g. if base were x**(1/2)*y**(1/3) then we should
-                # exponentiate by 6 and look for powers of x and y in the ratio
-                # of 2 to 3
-                qlcm = lcm([ratq(bi) for bi in Mul.make_args(bstart)])
-                if qlcm == 1:
-                    break  # we are done
-                b = bstart**qlcm
-                qlcm *= qstart
-                if all(ratq(bi) == 1 for bi in Mul.make_args(b)):
-                    last = True  # we are going to be done after this next pass
-            # this base no longer can find anything to join with and
-            # since it was longer than any other we are done with it
-            b, q = base
-            done.append((b, common_b.pop(base)*Rational(1, q)))
-
-        # update c_powers and get ready to continue with powsimp
-        c_powers = done
-        # there may be terms still in common_b that were bases that were
-        # identified as needing processing, so remove those, too
-        for (b, q), e in common_b.items():
-            if (b.is_Pow or b.func is exp) and \
-                    q is not S.One and not b.exp.is_Rational:
-                b, be = b.as_base_exp()
-                b = b**(be/q)
-            else:
-                b = root(b, q)
-            c_powers.append((b, e))
-        check = len(c_powers)
-        c_powers = dict(c_powers)
-        assert len(c_powers) == check  # there should have been no duplicates
-        # ==============================================================
-
-        # rebuild the expression
-        newexpr = expr.func(*(newexpr + [Pow(b, e) for b, e in c_powers.items()]))
-        if combine == 'exp':
-            return expr.func(newexpr, expr.func(*nc_part))
-        else:
-            return recurse(expr.func(*nc_part), combine='base') * \
-                recurse(newexpr, combine='base')
-
-    elif combine == 'base':
-
-        # Build c_powers and nc_part.  These must both be lists not
-        # dicts because exp's are not combined.
-        c_powers = []
-        nc_part = []
-        for term in expr.args:
-            if term.is_commutative:
-                c_powers.append(list(term.as_base_exp()))
-            else:
-                # This is the logic that combines bases that are
-                # different and non-commutative, but with equal and
-                # commutative exponents: A**x*B**x == (A*B)**x.
-                if nc_part:
-                    b1, e1 = nc_part[-1].as_base_exp()
-                    b2, e2 = term.as_base_exp()
-                    if (e1 == e2 and e2.is_commutative):
-                        nc_part[-1] = Pow(b1*b2, e1)
-                        continue
-                nc_part.append(term)
-
-        # Pull out numerical coefficients from exponent if assumptions allow
-        # e.g., 2**(2*x) => 4**x
-        for i in range(len(c_powers)):
-            b, e = c_powers[i]
-            if not (all(x.is_nonnegative for x in b.as_numer_denom()) or e.is_integer or force or b.is_polar):
-                continue
-            exp_c, exp_t = e.as_coeff_Mul(rational=True)
-            if exp_c is not S.One and exp_t is not S.One:
-                c_powers[i] = [Pow(b, exp_c), exp_t]
-
-        # Combine bases whenever they have the same exponent and
-        # assumptions allow
-        # first gather the potential bases under the common exponent
-        c_exp = defaultdict(list)
-        for b, e in c_powers:
-            if deep:
-                e = recurse(e)
-            c_exp[e].append(b)
-        del c_powers
-
-        # Merge back in the results of the above to form a new product
-        c_powers = defaultdict(list)
-        for e in c_exp:
-            bases = c_exp[e]
-
-            # calculate the new base for e
-
-            if len(bases) == 1:
-                new_base = bases[0]
-            elif e.is_integer or force:
-                new_base = expr.func(*bases)
-            else:
-                # see which ones can be joined
-                unk = []
-                nonneg = []
-                neg = []
-                for bi in bases:
-                    if bi.is_negative:
-                        neg.append(bi)
-                    elif bi.is_nonnegative:
-                        nonneg.append(bi)
-                    elif bi.is_polar:
-                        nonneg.append(
-                            bi)  # polar can be treated like non-negative
-                    else:
-                        unk.append(bi)
-                if len(unk) == 1 and not neg or len(neg) == 1 and not unk:
-                    # a single neg or a single unk can join the rest
-                    nonneg.extend(unk + neg)
-                    unk = neg = []
-                elif neg:
-                    # their negative signs cancel in groups of 2*q if we know
-                    # that e = p/q else we have to treat them as unknown
-                    israt = False
-                    if e.is_Rational:
-                        israt = True
-                    else:
-                        p, d = e.as_numer_denom()
-                        if p.is_integer and d.is_integer:
-                            israt = True
-                    if israt:
-                        neg = [-w for w in neg]
-                        unk.extend([S.NegativeOne]*len(neg))
-                    else:
-                        unk.extend(neg)
-                        neg = []
-                    del israt
-
-                # these shouldn't be joined
-                for b in unk:
-                    c_powers[b].append(e)
-                # here is a new joined base
-                new_base = expr.func(*(nonneg + neg))
-                # if there are positive parts they will just get separated
-                # again unless some change is made
-
-                def _terms(e):
-                    # return the number of terms of this expression
-                    # when multiplied out -- assuming no joining of terms
-                    if e.is_Add:
-                        return sum([_terms(ai) for ai in e.args])
-                    if e.is_Mul:
-                        return prod([_terms(mi) for mi in e.args])
-                    return 1
-                xnew_base = expand_mul(new_base, deep=False)
-                if len(Add.make_args(xnew_base)) < _terms(new_base):
-                    new_base = factor_terms(xnew_base)
-
-            c_powers[new_base].append(e)
-
-        # break out the powers from c_powers now
-        c_part = [Pow(b, ei) for b, e in c_powers.items() for ei in e]
-
-        # we're done
-        return expr.func(*(c_part + nc_part))
-
-    else:
-        raise ValueError("combine must be one of ('all', 'exp', 'base').")
+    eq = sympify(eq)
+    if iterable(eq):
+        f = type(eq)
+        eq = list(eq)
+        syms = set()
+        for e in eq:
+            syms = syms.union(e.atoms(Symbol))
+        reps = {}
+        for s in syms:
+            reps.update(dict((v, k) for k, v in posify(s)[1].items()))
+        for i, e in enumerate(eq):
+            eq[i] = e.subs(reps)
+        return f(eq), dict([(r, s) for s, r in reps.items()])
+
+    reps = dict([(s, Dummy(s.name, positive=True))
+                 for s in eq.free_symbols if s.is_positive is None])
+    eq = eq.subs(reps)
+    return eq, dict([(r, s) for s, r in reps.items()])
 
 
 def hypersimp(f, k):
@@ -3017,7 +714,6 @@ class _rf(Function):
                         return _rf(_a, b)*_rf(_a + b, c)/_rf(_a, c)
                     elif c < 0:
                         return _rf(_a, b)*_rf(_a + c, -c)/_rf(_a + b + c, -c)
-
 
 
 @timethis('combsimp')
@@ -3766,167 +1462,173 @@ def simplify(expr, ratio=1.7, measure=count_ops, fu=False):
     return expr
 
 
-def _real_to_rational(expr, tolerance=None):
-    """
-    Replace all reals in expr with rationals.
+def sum_simplify(s):
+    """Main function for Sum simplification"""
+    from sympy.concrete.summations import Sum
 
-    >>> from sympy import nsimplify
-    >>> from sympy.abc import x
+    terms = Add.make_args(s)
+    s_t = [] # Sum Terms
+    o_t = [] # Other Terms
 
-    >>> nsimplify(.76 + .1*x**.5, rational=True)
-    sqrt(x)/10 + 19/25
-
-    """
-    p = expr
-    reps = {}
-    reduce_num = None
-    if tolerance is not None and tolerance < 1:
-        reduce_num = ceiling(1/tolerance)
-    for float in p.atoms(Float):
-        key = float
-        if reduce_num is not None:
-            r = Rational(float).limit_denominator(reduce_num)
-        elif (tolerance is not None and tolerance >= 1 and
-                float.is_Integer is False):
-            r = Rational(tolerance*round(float/tolerance)
-                ).limit_denominator(int(tolerance))
-        else:
-            r = nsimplify(float, rational=False)
-            # e.g. log(3).n() -> log(3) instead of a Rational
-            if float and not r:
-                r = Rational(float)
-            elif not r.is_Rational:
-                if float < 0:
-                    float = -float
-                    d = Pow(10, int((mpmath.log(float)/mpmath.log(10))))
-                    r = -Rational(str(float/d))*d
-                elif float > 0:
-                    d = Pow(10, int((mpmath.log(float)/mpmath.log(10))))
-                    r = Rational(str(float/d))*d
+    for term in terms:
+        if isinstance(term, Mul):
+            constant = 1
+            other = 1
+            s = 0
+            n_sum_terms = 0
+            for j in range(len(term.args)):
+                if isinstance(term.args[j], Sum):
+                    s = term.args[j]
+                    n_sum_terms = n_sum_terms + 1
+                elif term.args[j].is_number == True:
+                    constant = constant * term.args[j]
                 else:
-                    r = Integer(0)
-        reps[key] = r
-    return p.subs(reps, simultaneous=True)
+                    other = other * term.args[j]
+            if other == 1 and n_sum_terms == 1:
+                # Insert the constant inside the Sum
+                s_t.append(Sum(constant * s.function, *s.limits))
+            elif other != 1 and n_sum_terms == 1:
+                o_t.append(other * Sum(constant * s.function, *s.limits))
+            else:
+                o_t.append(term)
+        elif isinstance(term, Sum):
+            s_t.append(term)
+        else:
+            o_t.append(term)
+
+    used = [False] * len(s_t)
+
+    for method in range(2):
+        for i, s_term1 in enumerate(s_t):
+            if not used[i]:
+                for j, s_term2 in enumerate(s_t):
+                    if not used[j] and i != j:
+                        temp = sum_add(s_term1, s_term2, method)
+                        if isinstance(temp, Sum):
+                            s_t[i] = temp
+                            s_term1 = s_t[i]
+                            used[j] = True
+
+    result = Add(*o_t)
+
+    for i, s_term in enumerate(s_t):
+        if not used[i]:
+            result = Add(result, s_term)
+
+    return result
 
 
-def nsimplify(expr, constants=[], tolerance=None, full=False, rational=None):
+def sum_add(self, other, method=0):
+    """Helper function for Sum simplification"""
+    from sympy.concrete.summations import Sum
+
+    if type(self) == type(other):
+        if method == 0:
+            if self.limits == other.limits:
+                return Sum(self.function + other.function, *self.limits)
+        elif method == 1:
+            if simplify(self.function - other.function) == 0:
+                if len(self.limits) == len(other.limits) == 1:
+                    i = self.limits[0][0]
+                    x1 = self.limits[0][1]
+                    y1 = self.limits[0][2]
+                    j = other.limits[0][0]
+                    x2 = other.limits[0][1]
+                    y2 = other.limits[0][2]
+
+                    if i == j:
+                        if x2 == y1 + 1:
+                            return Sum(self.function, (i, x1, y2))
+                        elif x1 == y2 + 1:
+                            return Sum(self.function, (i, x2, y1))
+
+    return Add(self, other)
+
+
+def product_simplify(s):
+    """Main function for Product simplification"""
+    from sympy.concrete.products import Product
+
+    terms = Mul.make_args(s)
+    p_t = [] # Product Terms
+    o_t = [] # Other Terms
+
+    for term in terms:
+        if isinstance(term, Product):
+            p_t.append(term)
+        else:
+            o_t.append(term)
+
+    used = [False] * len(p_t)
+
+    for method in range(2):
+        for i, p_term1 in enumerate(p_t):
+            if not used[i]:
+                for j, p_term2 in enumerate(p_t):
+                    if not used[j] and i != j:
+                        if isinstance(product_mul(p_term1, p_term2, method), Product):
+                            p_t[i] = product_mul(p_term1, p_term2, method)
+                            used[j] = True
+
+    result = Mul(*o_t)
+
+    for i, p_term in enumerate(p_t):
+        if not used[i]:
+            result = Mul(result, p_term)
+
+    return result
+
+
+def product_mul(self, other, method=0):
+    """Helper function for Product simplification"""
+    from sympy.concrete.products import Product
+
+    if type(self) == type(other):
+        if method == 0:
+            if self.limits == other.limits:
+                return Product(self.function * other.function, *self.limits)
+        elif method == 1:
+            if simplify(self.function - other.function) == 0:
+                if len(self.limits) == len(other.limits) == 1:
+                    i = self.limits[0][0]
+                    x1 = self.limits[0][1]
+                    y1 = self.limits[0][2]
+                    j = other.limits[0][0]
+                    x2 = other.limits[0][1]
+                    y2 = other.limits[0][2]
+
+                    if i == j:
+                        if x2 == y1 + 1:
+                            return Product(self.function, (i, x1, y2))
+                        elif x1 == y2 + 1:
+                            return Product(self.function, (i, x2, y1))
+
+    return Mul(self, other)
+
+
+def _nthroot_solve(p, n, prec):
     """
-    Find a simple representation for a number or, if there are free symbols or
-    if rational=True, then replace Floats with their Rational equivalents. If
-    no change is made and rational is not False then Floats will at least be
-    converted to Rationals.
-
-    For numerical expressions, a simple formula that numerically matches the
-    given numerical expression is sought (and the input should be possible
-    to evalf to a precision of at least 30 digits).
-
-    Optionally, a list of (rationally independent) constants to
-    include in the formula may be given.
-
-    A lower tolerance may be set to find less exact matches. If no tolerance
-    is given then the least precise value will set the tolerance (e.g. Floats
-    default to 15 digits of precision, so would be tolerance=10**-15).
-
-    With full=True, a more extensive search is performed
-    (this is useful to find simpler numbers when the tolerance
-    is set low).
-
-    Examples
-    ========
-
-    >>> from sympy import nsimplify, sqrt, GoldenRatio, exp, I, exp, pi
-    >>> nsimplify(4/(1+sqrt(5)), [GoldenRatio])
-    -2 + 2*GoldenRatio
-    >>> nsimplify((1/(exp(3*pi*I/5)+1)))
-    1/2 - I*sqrt(sqrt(5)/10 + 1/4)
-    >>> nsimplify(I**I, [pi])
-    exp(-pi/2)
-    >>> nsimplify(pi, tolerance=0.01)
-    22/7
-
-    See Also
-    ========
-    sympy.core.function.nfloat
-
+     helper function for ``nthroot``
+     It denests ``p**Rational(1, n)`` using its minimal polynomial
     """
-    try:
-        return sympify(as_int(expr))
-    except (TypeError, ValueError):
-        pass
-    expr = sympify(expr)
-    if rational or expr.free_symbols:
-        return _real_to_rational(expr, tolerance)
-
-    # SymPy's default tolerance for Rationals is 15; other numbers may have
-    # lower tolerances set, so use them to pick the largest tolerance if None
-    # was given
-    if tolerance is None:
-        tolerance = 10**-min([15] +
-             [mpmath.libmp.libmpf.prec_to_dps(n._prec)
-             for n in expr.atoms(Float)])
-    # XXX should prec be set independent of tolerance or should it be computed
-    # from tolerance?
-    prec = 30
-    bprec = int(prec*3.33)
-
-    constants_dict = {}
-    for constant in constants:
-        constant = sympify(constant)
-        v = constant.evalf(prec)
-        if not v.is_Float:
-            raise ValueError("constants must be real-valued")
-        constants_dict[str(constant)] = v._to_mpmath(bprec)
-
-    exprval = expr.evalf(prec, chop=True)
-    re, im = exprval.as_real_imag()
-
-    # safety check to make sure that this evaluated to a number
-    if not (re.is_Number and im.is_Number):
-        return expr
-
-    def nsimplify_real(x):
-        orig = mpmath.mp.dps
-        xv = x._to_mpmath(bprec)
-        try:
-            # We'll be happy with low precision if a simple fraction
-            if not (tolerance or full):
-                mpmath.mp.dps = 15
-                rat = mpmath.findpoly(xv, 1)
-                if rat is not None:
-                    return Rational(-int(rat[1]), int(rat[0]))
-            mpmath.mp.dps = prec
-            newexpr = mpmath.identify(xv, constants=constants_dict,
-                tol=tolerance, full=full)
-            if not newexpr:
-                raise ValueError
-            if full:
-                newexpr = newexpr[0]
-            expr = sympify(newexpr)
-            if x and not expr:  # don't let x become 0
-                raise ValueError
-            if expr.is_finite is False and not xv in [mpmath.inf, mpmath.ninf]:
-                raise ValueError
-            return expr
-        finally:
-            # even though there are returns above, this is executed
-            # before leaving
-            mpmath.mp.dps = orig
-    try:
-        if re:
-            re = nsimplify_real(re)
-        if im:
-            im = nsimplify_real(im)
-    except ValueError:
-        if rational is None:
-            return _real_to_rational(expr)
-        return expr
-
-    rv = re + im*S.ImaginaryUnit
-    # if there was a change or rational is explicitly not wanted
-    # return the value, else return the Rational representation
-    if rv != expr or rational is False:
-        return rv
-    return _real_to_rational(expr)
+    from sympy.polys.numberfields import _minimal_polynomial_sq
+    from sympy.solvers import solve
+    while n % 2 == 0:
+        p = sqrtdenest(sqrt(p))
+        n = n // 2
+    if n == 1:
+        return p
+    pn = p**Rational(1, n)
+    x = Symbol('x')
+    f = _minimal_polynomial_sq(p, n, x)
+    if f is None:
+        return None
+    sols = solve(f, x)
+    for sol in sols:
+        if abs(sol - pn).n() < 1./10**prec:
+            sol = sqrtdenest(sol)
+            if _mexpand(sol**n) == p:
+                return sol
 
 
 def logcombine(expr, force=False):
@@ -4088,890 +1790,224 @@ def bottom_up(rv, F, atoms=False, nonbasic=False):
     return rv
 
 
-def besselsimp(expr):
+def nthroot(expr, n, max_len=4, prec=15):
     """
-    Simplify bessel-type functions.
+    compute a real nth-root of a sum of surds
 
-    This routine tries to simplify bessel-type functions. Currently it only
-    works on the Bessel J and I functions, however. It works by looking at all
-    such functions in turn, and eliminating factors of "I" and "-1" (actually
-    their polar equivalents) in front of the argument. Then, functions of
-    half-integer order are rewritten using trigonometric functions and
-    functions of integer order (> 1) are rewritten using functions
-    of low order.  Finally, if the expression was changed, compute
-    factorization of the result with factor().
+    Parameters
+    ==========
 
-    >>> from sympy import besselj, besseli, besselsimp, polar_lift, I, S
-    >>> from sympy.abc import z, nu
-    >>> besselsimp(besselj(nu, z*polar_lift(-1)))
-    exp(I*pi*nu)*besselj(nu, z)
-    >>> besselsimp(besseli(nu, z*polar_lift(-I)))
-    exp(-I*pi*nu/2)*besselj(nu, z)
-    >>> besselsimp(besseli(S(-1)/2, z))
-    sqrt(2)*cosh(z)/(sqrt(pi)*sqrt(z))
-    >>> besselsimp(z*besseli(0, z) + z*(besseli(2, z))/2 + besseli(1, z))
-    3*z*besseli(0, z)/2
-    """
-    # TODO
-    # - better algorithm?
-    # - simplify (cos(pi*b)*besselj(b,z) - besselj(-b,z))/sin(pi*b) ...
-    # - use contiguity relations?
+    expr : sum of surds
+    n : integer
+    max_len : maximum number of surds passed as constants to ``nsimplify``
 
-    def replacer(fro, to, factors):
-        factors = set(factors)
+    Algorithm
+    =========
 
-        def repl(nu, z):
-            if factors.intersection(Mul.make_args(z)):
-                return to(nu, z)
-            return fro(nu, z)
-        return repl
-
-    def torewrite(fro, to):
-        def tofunc(nu, z):
-            return fro(nu, z).rewrite(to)
-        return tofunc
-
-    def tominus(fro):
-        def tofunc(nu, z):
-            return exp(I*pi*nu)*fro(nu, exp_polar(-I*pi)*z)
-        return tofunc
-
-    orig_expr = expr
-
-    ifactors = [I, exp_polar(I*pi/2), exp_polar(-I*pi/2)]
-    expr = expr.replace(
-        besselj, replacer(besselj,
-        torewrite(besselj, besseli), ifactors))
-    expr = expr.replace(
-        besseli, replacer(besseli,
-        torewrite(besseli, besselj), ifactors))
-
-    minusfactors = [-1, exp_polar(I*pi)]
-    expr = expr.replace(
-        besselj, replacer(besselj, tominus(besselj), minusfactors))
-    expr = expr.replace(
-        besseli, replacer(besseli, tominus(besseli), minusfactors))
-
-    z0 = Dummy('z')
-
-    def expander(fro):
-        def repl(nu, z):
-            if (nu % 1) == S(1)/2:
-                return exptrigsimp(trigsimp(unpolarify(
-                        fro(nu, z0).rewrite(besselj).rewrite(jn).expand(
-                            func=True)).subs(z0, z)))
-            elif nu.is_Integer and nu > 1:
-                return fro(nu, z).expand(func=True)
-            return fro(nu, z)
-        return repl
-
-    expr = expr.replace(besselj, expander(besselj))
-    expr = expr.replace(bessely, expander(bessely))
-    expr = expr.replace(besseli, expander(besseli))
-    expr = expr.replace(besselk, expander(besselk))
-
-    if expr != orig_expr:
-        expr = expr.factor()
-
-    return expr
-
-def exptrigsimp(expr, simplify=True):
-    """
-    Simplifies exponential / trigonometric / hyperbolic functions.
-    When ``simplify`` is True (default) the expression obtained after the
-    simplification step will be then be passed through simplify to
-    precondition it so the final transformations will be applied.
+    First ``nsimplify`` is used to get a candidate root; if it is not a
+    root the minimal polynomial is computed; the answer is one of its
+    roots.
 
     Examples
     ========
 
-    >>> from sympy import exptrigsimp, exp, cosh, sinh
-    >>> from sympy.abc import z
+    >>> from sympy.simplify.simplify import nthroot
+    >>> from sympy import Rational, sqrt
+    >>> nthroot(90 + 34*sqrt(7), 3)
+    sqrt(7) + 3
 
-    >>> exptrigsimp(exp(z) + exp(-z))
-    2*cosh(z)
-    >>> exptrigsimp(cosh(z) - sinh(z))
-    exp(-z)
     """
-    from sympy.simplify.fu import hyper_as_trig, TR2i
-
-    def exp_trig(e):
-        # select the better of e, and e rewritten in terms of exp or trig
-        # functions
-        choices = [e]
-        if e.has(*_trigs):
-            choices.append(e.rewrite(exp))
-        choices.append(e.rewrite(cos))
-        return min(*choices, key=count_ops)
-    newexpr = bottom_up(expr, exp_trig)
-
-    if simplify:
-        newexpr = newexpr.simplify()
-
-    # conversion from exp to hyperbolic
-    ex = newexpr.atoms(exp, S.Exp1)
-    ex = [ei for ei in ex if 1/ei not in ex]
-    ## sinh and cosh
-    for ei in ex:
-        e2 = ei**-2
-        if e2 in ex:
-            a = e2.args[0]/2 if not e2 is S.Exp1 else S.Half
-            newexpr = newexpr.subs((e2 + 1)*ei, 2*cosh(a))
-            newexpr = newexpr.subs((e2 - 1)*ei, 2*sinh(a))
-    ## exp ratios to tan and tanh
-    for ei in ex:
-        n, d = ei - 1, ei + 1
-        et = n/d
-        etinv = d/n  # not 1/et or else recursion errors arise
-        a = ei.args[0] if ei.func is exp else S.One
-        if a.is_Mul or a is S.ImaginaryUnit:
-            c = a.as_coefficient(I)
-            if c:
-                t = S.ImaginaryUnit*tan(c/2)
-                newexpr = newexpr.subs(etinv, 1/t)
-                newexpr = newexpr.subs(et, t)
-                continue
-        t = tanh(a/2)
-        newexpr = newexpr.subs(etinv, 1/t)
-        newexpr = newexpr.subs(et, t)
-
-    # sin/cos and sinh/cosh ratios to tan and tanh, respectively
-    if newexpr.has(HyperbolicFunction):
-        e, f = hyper_as_trig(newexpr)
-        newexpr = f(TR2i(e))
-    if newexpr.has(TrigonometricFunction):
-        newexpr = TR2i(newexpr)
-
-    # can we ever generate an I where there was none previously?
-    if not (newexpr.has(I) and not expr.has(I)):
-        expr = newexpr
+    expr = sympify(expr)
+    n = sympify(n)
+    p = expr**Rational(1, n)
+    if not n.is_integer:
+        return p
+    if not _is_sum_surds(expr):
+        return p
+    surds = []
+    coeff_muls = [x.as_coeff_Mul() for x in expr.args]
+    for x, y in coeff_muls:
+        if not x.is_rational:
+            return p
+        if y is S.One:
+            continue
+        if not (y.is_Pow and y.exp == S.Half and y.base.is_integer):
+            return p
+        surds.append(y)
+    surds.sort()
+    surds = surds[:max_len]
+    if expr < 0 and n % 2 == 1:
+        p = (-expr)**Rational(1, n)
+        a = nsimplify(p, constants=surds)
+        res = a if _mexpand(a**n) == _mexpand(-expr) else p
+        return -res
+    a = nsimplify(p, constants=surds)
+    if _mexpand(a) is not _mexpand(p) and _mexpand(a**n) == _mexpand(expr):
+        return _mexpand(a)
+    expr = _nthroot_solve(expr, n, prec)
+    if expr is None:
+        return p
     return expr
 
 
-def _is_Expr(e):
-    """_eapply helper to tell whether ``e`` and all its args
-    are Exprs."""
-    if not isinstance(e, Expr):
-        return False
-    return all(_is_Expr(i) for i in e.args)
+def nsimplify(expr, constants=[], tolerance=None, full=False, rational=None):
+    """
+    Find a simple representation for a number or, if there are free symbols or
+    if rational=True, then replace Floats with their Rational equivalents. If
+    no change is made and rational is not False then Floats will at least be
+    converted to Rationals.
 
+    For numerical expressions, a simple formula that numerically matches the
+    given numerical expression is sought (and the input should be possible
+    to evalf to a precision of at least 30 digits).
 
-def _eapply(func, e, cond=None):
-    """Apply ``func`` to ``e`` if all args are Exprs else only
-    apply it to those args that *are* Exprs."""
-    if not isinstance(e, Expr):
-        return e
-    if _is_Expr(e) or not e.args:
-        return func(e)
-    return e.func(*[
-        _eapply(func, ei) if (cond is None or cond(ei)) else ei
-        for ei in e.args])
+    Optionally, a list of (rationally independent) constants to
+    include in the formula may be given.
 
+    A lower tolerance may be set to find less exact matches. If no tolerance
+    is given then the least precise value will set the tolerance (e.g. Floats
+    default to 15 digits of precision, so would be tolerance=10**-15).
 
-def futrig(e, **kwargs):
-    """Return simplified ``e`` using Fu-like transformations.
-    This is not the "Fu" algorithm. This is called by default
-    from ``trigsimp``. By default, hyperbolics subexpressions
-    will be simplified, but this can be disabled by setting
-    ``hyper=False``.
+    With full=True, a more extensive search is performed
+    (this is useful to find simpler numbers when the tolerance
+    is set low).
 
     Examples
     ========
 
-    >>> from sympy import trigsimp, tan, sinh, tanh
-    >>> from sympy.simplify.simplify import futrig
-    >>> from sympy.abc import x
-    >>> trigsimp(1/tan(x)**2)
-    tan(x)**(-2)
+    >>> from sympy import nsimplify, sqrt, GoldenRatio, exp, I, exp, pi
+    >>> nsimplify(4/(1+sqrt(5)), [GoldenRatio])
+    -2 + 2*GoldenRatio
+    >>> nsimplify((1/(exp(3*pi*I/5)+1)))
+    1/2 - I*sqrt(sqrt(5)/10 + 1/4)
+    >>> nsimplify(I**I, [pi])
+    exp(-pi/2)
+    >>> nsimplify(pi, tolerance=0.01)
+    22/7
 
-    >>> futrig(sinh(x)/tanh(x))
-    cosh(x)
-
-    """
-    from sympy.simplify.fu import hyper_as_trig
-
-    e = sympify(e)
-
-    if not isinstance(e, Basic):
-        return e
-
-    if not e.args:
-        return e
-
-    old = e
-    e = bottom_up(e, lambda x: _futrig(x, **kwargs))
-
-    if kwargs.pop('hyper', True) and e.has(HyperbolicFunction):
-        e, f = hyper_as_trig(e)
-        e = f(_futrig(e))
-
-    if e != old and e.is_Mul and e.args[0].is_Rational:
-        # redistribute leading coeff on 2-arg Add
-        e = Mul(*e.as_coeff_Mul())
-    return e
-
-
-def _futrig(e, **kwargs):
-    """Helper for futrig."""
-    from sympy.simplify.fu import (
-        TR1, TR2, TR3, TR2i, TR10, L, TR10i,
-        TR8, TR6, TR15, TR16, TR111, TR5, TRmorrie, TR11, TR14, TR22,
-        TR12)
-    from sympy.core.compatibility import _nodes
-
-    if not e.has(TrigonometricFunction):
-        return e
-
-    if e.is_Mul:
-        coeff, e = e.as_independent(TrigonometricFunction)
-    else:
-        coeff = S.One
-
-    Lops = lambda x: (L(x), x.count_ops(), _nodes(x), len(x.args), x.is_Add)
-    trigs = lambda x: x.has(TrigonometricFunction)
-
-    tree = [identity,
-        (
-        TR3,  # canonical angles
-        TR1,  # sec-csc -> cos-sin
-        TR12,  # expand tan of sum
-        lambda x: _eapply(factor, x, trigs),
-        TR2,  # tan-cot -> sin-cos
-        [identity, lambda x: _eapply(_mexpand, x, trigs)],
-        TR2i,  # sin-cos ratio -> tan
-        lambda x: _eapply(lambda i: factor(i.normal()), x, trigs),
-        TR14,  # factored identities
-        TR5,  # sin-pow -> cos_pow
-        TR10,  # sin-cos of sums -> sin-cos prod
-        TR11, TR6, # reduce double angles and rewrite cos pows
-        lambda x: _eapply(factor, x, trigs),
-        TR14,  # factored powers of identities
-        [identity, lambda x: _eapply(_mexpand, x, trigs)],
-        TRmorrie,
-        TR10i,  # sin-cos products > sin-cos of sums
-        [identity, TR8],  # sin-cos products -> sin-cos of sums
-        [identity, lambda x: TR2i(TR2(x))],  # tan -> sin-cos -> tan
-        [
-            lambda x: _eapply(expand_mul, TR5(x), trigs),
-            lambda x: _eapply(
-                expand_mul, TR15(x), trigs)], # pos/neg powers of sin
-        [
-            lambda x:  _eapply(expand_mul, TR6(x), trigs),
-            lambda x:  _eapply(
-                expand_mul, TR16(x), trigs)], # pos/neg powers of cos
-        TR111,  # tan, sin, cos to neg power -> cot, csc, sec
-        [identity, TR2i],  # sin-cos ratio to tan
-        [identity, lambda x: _eapply(
-            expand_mul, TR22(x), trigs)],  # tan-cot to sec-csc
-        TR1, TR2, TR2i,
-        [identity, lambda x: _eapply(
-            factor_terms, TR12(x), trigs)],  # expand tan of sum
-        )]
-    e = greedy(tree, objective=Lops)(e)
-
-    return coeff*e
-
-
-def sum_simplify(s):
-    """Main function for Sum simplification"""
-    from sympy.concrete.summations import Sum
-
-    terms = Add.make_args(s)
-    s_t = [] # Sum Terms
-    o_t = [] # Other Terms
-
-    for term in terms:
-        if isinstance(term, Mul):
-            constant = 1
-            other = 1
-            s = 0
-            n_sum_terms = 0
-            for j in range(len(term.args)):
-                if isinstance(term.args[j], Sum):
-                    s = term.args[j]
-                    n_sum_terms = n_sum_terms + 1
-                elif term.args[j].is_number == True:
-                    constant = constant * term.args[j]
-                else:
-                    other = other * term.args[j]
-            if other == 1 and n_sum_terms == 1:
-                # Insert the constant inside the Sum
-                s_t.append(Sum(constant * s.function, *s.limits))
-            elif other != 1 and n_sum_terms == 1:
-                o_t.append(other * Sum(constant * s.function, *s.limits))
-            else:
-                o_t.append(term)
-        elif isinstance(term, Sum):
-            s_t.append(term)
-        else:
-            o_t.append(term)
-
-    used = [False] * len(s_t)
-
-    for method in range(2):
-        for i, s_term1 in enumerate(s_t):
-            if not used[i]:
-                for j, s_term2 in enumerate(s_t):
-                    if not used[j] and i != j:
-                        temp = sum_add(s_term1, s_term2, method)
-                        if isinstance(temp, Sum):
-                            s_t[i] = temp
-                            s_term1 = s_t[i]
-                            used[j] = True
-
-    result = Add(*o_t)
-
-    for i, s_term in enumerate(s_t):
-        if not used[i]:
-            result = Add(result, s_term)
-
-    return result
-
-
-def sum_add(self, other, method=0):
-    """Helper function for Sum simplification"""
-    from sympy.concrete.summations import Sum
-
-    if type(self) == type(other):
-        if method == 0:
-            if self.limits == other.limits:
-                return Sum(self.function + other.function, *self.limits)
-        elif method == 1:
-            if simplify(self.function - other.function) == 0:
-                if len(self.limits) == len(other.limits) == 1:
-                    i = self.limits[0][0]
-                    x1 = self.limits[0][1]
-                    y1 = self.limits[0][2]
-                    j = other.limits[0][0]
-                    x2 = other.limits[0][1]
-                    y2 = other.limits[0][2]
-
-                    if i == j:
-                        if x2 == y1 + 1:
-                            return Sum(self.function, (i, x1, y2))
-                        elif x1 == y2 + 1:
-                            return Sum(self.function, (i, x2, y1))
-
-    return Add(self, other)
-
-
-def product_simplify(s):
-    """Main function for Product simplification"""
-    from sympy.concrete.products import Product
-
-    terms = Mul.make_args(s)
-    p_t = [] # Product Terms
-    o_t = [] # Other Terms
-
-    for term in terms:
-        if isinstance(term, Product):
-            p_t.append(term)
-        else:
-            o_t.append(term)
-
-    used = [False] * len(p_t)
-
-    for method in range(2):
-        for i, p_term1 in enumerate(p_t):
-            if not used[i]:
-                for j, p_term2 in enumerate(p_t):
-                    if not used[j] and i != j:
-                        if isinstance(product_mul(p_term1, p_term2, method), Product):
-                            p_t[i] = product_mul(p_term1, p_term2, method)
-                            used[j] = True
-
-    result = Mul(*o_t)
-
-    for i, p_term in enumerate(p_t):
-        if not used[i]:
-            result = Mul(result, p_term)
-
-    return result
-
-
-def product_mul(self, other, method=0):
-    """Helper function for Product simplification"""
-    from sympy.concrete.products import Product
-
-    if type(self) == type(other):
-        if method == 0:
-            if self.limits == other.limits:
-                return Product(self.function * other.function, *self.limits)
-        elif method == 1:
-            if simplify(self.function - other.function) == 0:
-                if len(self.limits) == len(other.limits) == 1:
-                    i = self.limits[0][0]
-                    x1 = self.limits[0][1]
-                    y1 = self.limits[0][2]
-                    j = other.limits[0][0]
-                    x2 = other.limits[0][1]
-                    y2 = other.limits[0][2]
-
-                    if i == j:
-                        if x2 == y1 + 1:
-                            return Product(self.function, (i, x1, y2))
-                        elif x1 == y2 + 1:
-                            return Product(self.function, (i, x2, y1))
-
-    return Mul(self, other)
-
-#-------------------- the old trigsimp routines ---------------------
-
-def trigsimp_old(expr, **opts):
-    """
-    reduces expression by using known trig identities
-
-    Notes
-    =====
-
-    deep:
-    - Apply trigsimp inside all objects with arguments
-
-    recursive:
-    - Use common subexpression elimination (cse()) and apply
-    trigsimp recursively (this is quite expensive if the
-    expression is large)
-
-    method:
-    - Determine the method to use. Valid choices are 'matching' (default),
-    'groebner', 'combined', 'fu' and 'futrig'. If 'matching', simplify the
-    expression recursively by pattern matching. If 'groebner', apply an
-    experimental groebner basis algorithm. In this case further options
-    are forwarded to ``trigsimp_groebner``, please refer to its docstring.
-    If 'combined', first run the groebner basis algorithm with small
-    default parameters, then run the 'matching' algorithm. 'fu' runs the
-    collection of trigonometric transformations described by Fu, et al.
-    (see the `fu` docstring) while `futrig` runs a subset of Fu-transforms
-    that mimic the behavior of `trigsimp`.
-
-    compare:
-    - show input and output from `trigsimp` and `futrig` when different,
-    but returns the `trigsimp` value.
-
-    Examples
+    See Also
     ========
-
-    >>> from sympy import trigsimp, sin, cos, log, cosh, sinh, tan, cot
-    >>> from sympy.abc import x, y
-    >>> e = 2*sin(x)**2 + 2*cos(x)**2
-    >>> trigsimp(e, old=True)
-    2
-    >>> trigsimp(log(e), old=True)
-    log(2*sin(x)**2 + 2*cos(x)**2)
-    >>> trigsimp(log(e), deep=True, old=True)
-    log(2)
-
-    Using `method="groebner"` (or `"combined"`) can sometimes lead to a lot
-    more simplification:
-
-    >>> e = (-sin(x) + 1)/cos(x) + cos(x)/(-sin(x) + 1)
-    >>> trigsimp(e, old=True)
-    (-sin(x) + 1)/cos(x) - cos(x)/(sin(x) - 1)
-    >>> trigsimp(e, method="groebner", old=True)
-    2/cos(x)
-
-    >>> trigsimp(1/cot(x)**2, compare=True, old=True)
-          futrig: tan(x)**2
-    cot(x)**(-2)
+    sympy.core.function.nfloat
 
     """
-    old = expr
-    first = opts.pop('first', True)
-    if first:
-        if not expr.has(*_trigs):
-            return expr
-
-        trigsyms = set().union(*[t.free_symbols for t in expr.atoms(*_trigs)])
-        if len(trigsyms) > 1:
-            d = separatevars(expr)
-            if d.is_Mul:
-                d = separatevars(d, dict=True) or d
-            if isinstance(d, dict):
-                expr = 1
-                for k, v in d.items():
-                    # remove hollow factoring
-                    was = v
-                    v = expand_mul(v)
-                    opts['first'] = False
-                    vnew = trigsimp(v, **opts)
-                    if vnew == v:
-                        vnew = was
-                    expr *= vnew
-                old = expr
-            else:
-                if d.is_Add:
-                    for s in trigsyms:
-                        r, e = expr.as_independent(s)
-                        if r:
-                            opts['first'] = False
-                            expr = r + trigsimp(e, **opts)
-                            if not expr.is_Add:
-                                break
-                    old = expr
-
-    recursive = opts.pop('recursive', False)
-    deep = opts.pop('deep', False)
-    method = opts.pop('method', 'matching')
-
-    def groebnersimp(ex, deep, **opts):
-        def traverse(e):
-            if e.is_Atom:
-                return e
-            args = [traverse(x) for x in e.args]
-            if e.is_Function or e.is_Pow:
-                args = [trigsimp_groebner(x, **opts) for x in args]
-            return e.func(*args)
-        if deep:
-            ex = traverse(ex)
-        return trigsimp_groebner(ex, **opts)
-
-    trigsimpfunc = {
-        'matching': (lambda x, d: _trigsimp(x, d)),
-        'groebner': (lambda x, d: groebnersimp(x, d, **opts)),
-        'combined': (lambda x, d: _trigsimp(groebnersimp(x,
-                                       d, polynomial=True, hints=[2, tan]),
-                                   d))
-                   }[method]
-
-    if recursive:
-        w, g = cse(expr)
-        g = trigsimpfunc(g[0], deep)
-
-        for sub in reversed(w):
-            g = g.subs(sub[0], sub[1])
-            g = trigsimpfunc(g, deep)
-        result = g
-    else:
-        result = trigsimpfunc(expr, deep)
-
-    if opts.get('compare', False):
-        f = futrig(old)
-        if f != result:
-            print('\tfutrig:', f)
-
-    return result
-
-
-def _dotrig(a, b):
-    """Helper to tell whether ``a`` and ``b`` have the same sorts
-    of symbols in them -- no need to test hyperbolic patterns against
-    expressions that have no hyperbolics in them."""
-    return a.func == b.func and (
-        a.has(TrigonometricFunction) and b.has(TrigonometricFunction) or
-        a.has(HyperbolicFunction) and b.has(HyperbolicFunction))
-
-
-_trigpat = None
-def _trigpats():
-    global _trigpat
-    a, b, c = symbols('a b c', cls=Wild)
-    d = Wild('d', commutative=False)
-
-    # for the simplifications like sinh/cosh -> tanh:
-    # DO NOT REORDER THE FIRST 14 since these are assumed to be in this
-    # order in _match_div_rewrite.
-    matchers_division = (
-        (a*sin(b)**c/cos(b)**c, a*tan(b)**c, sin(b), cos(b)),
-        (a*tan(b)**c*cos(b)**c, a*sin(b)**c, sin(b), cos(b)),
-        (a*cot(b)**c*sin(b)**c, a*cos(b)**c, sin(b), cos(b)),
-        (a*tan(b)**c/sin(b)**c, a/cos(b)**c, sin(b), cos(b)),
-        (a*cot(b)**c/cos(b)**c, a/sin(b)**c, sin(b), cos(b)),
-        (a*cot(b)**c*tan(b)**c, a, sin(b), cos(b)),
-        (a*(cos(b) + 1)**c*(cos(b) - 1)**c,
-            a*(-sin(b)**2)**c, cos(b) + 1, cos(b) - 1),
-        (a*(sin(b) + 1)**c*(sin(b) - 1)**c,
-            a*(-cos(b)**2)**c, sin(b) + 1, sin(b) - 1),
-
-        (a*sinh(b)**c/cosh(b)**c, a*tanh(b)**c, S.One, S.One),
-        (a*tanh(b)**c*cosh(b)**c, a*sinh(b)**c, S.One, S.One),
-        (a*coth(b)**c*sinh(b)**c, a*cosh(b)**c, S.One, S.One),
-        (a*tanh(b)**c/sinh(b)**c, a/cosh(b)**c, S.One, S.One),
-        (a*coth(b)**c/cosh(b)**c, a/sinh(b)**c, S.One, S.One),
-        (a*coth(b)**c*tanh(b)**c, a, S.One, S.One),
-
-        (c*(tanh(a) + tanh(b))/(1 + tanh(a)*tanh(b)),
-            tanh(a + b)*c, S.One, S.One),
-    )
-
-    matchers_add = (
-        (c*sin(a)*cos(b) + c*cos(a)*sin(b) + d, sin(a + b)*c + d),
-        (c*cos(a)*cos(b) - c*sin(a)*sin(b) + d, cos(a + b)*c + d),
-        (c*sin(a)*cos(b) - c*cos(a)*sin(b) + d, sin(a - b)*c + d),
-        (c*cos(a)*cos(b) + c*sin(a)*sin(b) + d, cos(a - b)*c + d),
-        (c*sinh(a)*cosh(b) + c*sinh(b)*cosh(a) + d, sinh(a + b)*c + d),
-        (c*cosh(a)*cosh(b) + c*sinh(a)*sinh(b) + d, cosh(a + b)*c + d),
-    )
-
-    # for cos(x)**2 + sin(x)**2 -> 1
-    matchers_identity = (
-        (a*sin(b)**2, a - a*cos(b)**2),
-        (a*tan(b)**2, a*(1/cos(b))**2 - a),
-        (a*cot(b)**2, a*(1/sin(b))**2 - a),
-        (a*sin(b + c), a*(sin(b)*cos(c) + sin(c)*cos(b))),
-        (a*cos(b + c), a*(cos(b)*cos(c) - sin(b)*sin(c))),
-        (a*tan(b + c), a*((tan(b) + tan(c))/(1 - tan(b)*tan(c)))),
-
-        (a*sinh(b)**2, a*cosh(b)**2 - a),
-        (a*tanh(b)**2, a - a*(1/cosh(b))**2),
-        (a*coth(b)**2, a + a*(1/sinh(b))**2),
-        (a*sinh(b + c), a*(sinh(b)*cosh(c) + sinh(c)*cosh(b))),
-        (a*cosh(b + c), a*(cosh(b)*cosh(c) + sinh(b)*sinh(c))),
-        (a*tanh(b + c), a*((tanh(b) + tanh(c))/(1 + tanh(b)*tanh(c)))),
-
-    )
-
-    # Reduce any lingering artifacts, such as sin(x)**2 changing
-    # to 1-cos(x)**2 when sin(x)**2 was "simpler"
-    artifacts = (
-        (a - a*cos(b)**2 + c, a*sin(b)**2 + c, cos),
-        (a - a*(1/cos(b))**2 + c, -a*tan(b)**2 + c, cos),
-        (a - a*(1/sin(b))**2 + c, -a*cot(b)**2 + c, sin),
-
-        (a - a*cosh(b)**2 + c, -a*sinh(b)**2 + c, cosh),
-        (a - a*(1/cosh(b))**2 + c, a*tanh(b)**2 + c, cosh),
-        (a + a*(1/sinh(b))**2 + c, a*coth(b)**2 + c, sinh),
-
-        # same as above but with noncommutative prefactor
-        (a*d - a*d*cos(b)**2 + c, a*d*sin(b)**2 + c, cos),
-        (a*d - a*d*(1/cos(b))**2 + c, -a*d*tan(b)**2 + c, cos),
-        (a*d - a*d*(1/sin(b))**2 + c, -a*d*cot(b)**2 + c, sin),
-
-        (a*d - a*d*cosh(b)**2 + c, -a*d*sinh(b)**2 + c, cosh),
-        (a*d - a*d*(1/cosh(b))**2 + c, a*d*tanh(b)**2 + c, cosh),
-        (a*d + a*d*(1/sinh(b))**2 + c, a*d*coth(b)**2 + c, sinh),
-    )
-
-    _trigpat = (a, b, c, d, matchers_division, matchers_add,
-        matchers_identity, artifacts)
-    return _trigpat
-
-
-def _replace_mul_fpowxgpow(expr, f, g, rexp, h, rexph):
-    """Helper for _match_div_rewrite.
-
-    Replace f(b_)**c_*g(b_)**(rexp(c_)) with h(b)**rexph(c) if f(b_)
-    and g(b_) are both positive or if c_ is an integer.
-    """
-    # assert expr.is_Mul and expr.is_commutative and f != g
-    fargs = defaultdict(int)
-    gargs = defaultdict(int)
-    args = []
-    for x in expr.args:
-        if x.is_Pow or x.func in (f, g):
-            b, e = x.as_base_exp()
-            if b.is_positive or e.is_integer:
-                if b.func == f:
-                    fargs[b.args[0]] += e
-                    continue
-                elif b.func == g:
-                    gargs[b.args[0]] += e
-                    continue
-        args.append(x)
-    common = set(fargs) & set(gargs)
-    hit = False
-    while common:
-        key = common.pop()
-        fe = fargs.pop(key)
-        ge = gargs.pop(key)
-        if fe == rexp(ge):
-            args.append(h(key)**rexph(fe))
-            hit = True
-        else:
-            fargs[key] = fe
-            gargs[key] = ge
-    if not hit:
-        return expr
-    while fargs:
-        key, e = fargs.popitem()
-        args.append(f(key)**e)
-    while gargs:
-        key, e = gargs.popitem()
-        args.append(g(key)**e)
-    return Mul(*args)
-
-
-_idn = lambda x: x
-_midn = lambda x: -x
-_one = lambda x: S.One
-
-def _match_div_rewrite(expr, i):
-    """helper for __trigsimp"""
-    if i == 0:
-        expr = _replace_mul_fpowxgpow(expr, sin, cos,
-            _midn, tan, _idn)
-    elif i == 1:
-        expr = _replace_mul_fpowxgpow(expr, tan, cos,
-            _idn, sin, _idn)
-    elif i == 2:
-        expr = _replace_mul_fpowxgpow(expr, cot, sin,
-            _idn, cos, _idn)
-    elif i == 3:
-        expr = _replace_mul_fpowxgpow(expr, tan, sin,
-            _midn, cos, _midn)
-    elif i == 4:
-        expr = _replace_mul_fpowxgpow(expr, cot, cos,
-            _midn, sin, _midn)
-    elif i == 5:
-        expr = _replace_mul_fpowxgpow(expr, cot, tan,
-            _idn, _one, _idn)
-    # i in (6, 7) is skipped
-    elif i == 8:
-        expr = _replace_mul_fpowxgpow(expr, sinh, cosh,
-            _midn, tanh, _idn)
-    elif i == 9:
-        expr = _replace_mul_fpowxgpow(expr, tanh, cosh,
-            _idn, sinh, _idn)
-    elif i == 10:
-        expr = _replace_mul_fpowxgpow(expr, coth, sinh,
-            _idn, cosh, _idn)
-    elif i == 11:
-        expr = _replace_mul_fpowxgpow(expr, tanh, sinh,
-            _midn, cosh, _midn)
-    elif i == 12:
-        expr = _replace_mul_fpowxgpow(expr, coth, cosh,
-            _midn, sinh, _midn)
-    elif i == 13:
-        expr = _replace_mul_fpowxgpow(expr, coth, tanh,
-            _idn, _one, _idn)
-    else:
-        return None
-    return expr
-
-
-def _trigsimp(expr, deep=False):
-    # protect the cache from non-trig patterns; we only allow
-    # trig patterns to enter the cache
-    if expr.has(*_trigs):
-        return __trigsimp(expr, deep)
-    return expr
-
-
-@cacheit
-def __trigsimp(expr, deep=False):
-    """recursive helper for trigsimp"""
-    from sympy.simplify.fu import TR10i
-
-    if _trigpat is None:
-        _trigpats()
-    a, b, c, d, matchers_division, matchers_add, \
-    matchers_identity, artifacts = _trigpat
-
-    if expr.is_Mul:
-        # do some simplifications like sin/cos -> tan:
-        if not expr.is_commutative:
-            com, nc = expr.args_cnc()
-            expr = _trigsimp(Mul._from_args(com), deep)*Mul._from_args(nc)
-        else:
-            for i, (pattern, simp, ok1, ok2) in enumerate(matchers_division):
-                if not _dotrig(expr, pattern):
-                    continue
-
-                newexpr = _match_div_rewrite(expr, i)
-                if newexpr is not None:
-                    if newexpr != expr:
-                        expr = newexpr
-                        break
-                    else:
-                        continue
-
-                # use SymPy matching instead
-                res = expr.match(pattern)
-                if res and res.get(c, 0):
-                    if not res[c].is_integer:
-                        ok = ok1.subs(res)
-                        if not ok.is_positive:
-                            continue
-                        ok = ok2.subs(res)
-                        if not ok.is_positive:
-                            continue
-                    # if "a" contains any of trig or hyperbolic funcs with
-                    # argument "b" then skip the simplification
-                    if any(w.args[0] == res[b] for w in res[a].atoms(
-                            TrigonometricFunction, HyperbolicFunction)):
-                        continue
-                    # simplify and finish:
-                    expr = simp.subs(res)
-                    break  # process below
-
-    if expr.is_Add:
-        args = []
-        for term in expr.args:
-            if not term.is_commutative:
-                com, nc = term.args_cnc()
-                nc = Mul._from_args(nc)
-                term = Mul._from_args(com)
-            else:
-                nc = S.One
-            term = _trigsimp(term, deep)
-            for pattern, result in matchers_identity:
-                res = term.match(pattern)
-                if res is not None:
-                    term = result.subs(res)
-                    break
-            args.append(term*nc)
-        if args != expr.args:
-            expr = Add(*args)
-            expr = min(expr, expand(expr), key=count_ops)
-        if expr.is_Add:
-            for pattern, result in matchers_add:
-                if not _dotrig(expr, pattern):
-                    continue
-                expr = TR10i(expr)
-                if expr.has(HyperbolicFunction):
-                    res = expr.match(pattern)
-                    # if "d" contains any trig or hyperbolic funcs with
-                    # argument "a" or "b" then skip the simplification;
-                    # this isn't perfect -- see tests
-                    if res is None or not (a in res and b in res) or any(
-                        w.args[0] in (res[a], res[b]) for w in res[d].atoms(
-                            TrigonometricFunction, HyperbolicFunction)):
-                        continue
-                    expr = result.subs(res)
-                    break
-
-        # Reduce any lingering artifacts, such as sin(x)**2 changing
-        # to 1 - cos(x)**2 when sin(x)**2 was "simpler"
-        for pattern, result, ex in artifacts:
-            if not _dotrig(expr, pattern):
-                continue
-            # Substitute a new wild that excludes some function(s)
-            # to help influence a better match. This is because
-            # sometimes, for example, 'a' would match sec(x)**2
-            a_t = Wild('a', exclude=[ex])
-            pattern = pattern.subs(a, a_t)
-            result = result.subs(a, a_t)
-
-            m = expr.match(pattern)
-            was = None
-            while m and was != expr:
-                was = expr
-                if m[a_t] == 0 or \
-                        -m[a_t] in m[c].args or m[a_t] + m[c] == 0:
-                    break
-                if d in m and m[a_t]*m[d] + m[c] == 0:
-                    break
-                expr = result.subs(m)
-                m = expr.match(pattern)
-                m.setdefault(c, S.Zero)
-
-    elif expr.is_Mul or expr.is_Pow or deep and expr.args:
-        expr = expr.func(*[_trigsimp(a, deep) for a in expr.args])
-
     try:
-        if not expr.has(*_trigs):
-            raise TypeError
-        e = expr.atoms(exp)
-        new = expr.rewrite(exp, deep=deep)
-        if new == e:
-            raise TypeError
-        fnew = factor(new)
-        if fnew != new:
-            new = sorted([new, factor(new)], key=count_ops)[0]
-        # if all exp that were introduced disappeared then accept it
-        if not (new.atoms(exp) - e):
-            expr = new
-    except TypeError:
+        return sympify(as_int(expr))
+    except (TypeError, ValueError):
         pass
+    expr = sympify(expr)
+    if rational or expr.free_symbols:
+        return _real_to_rational(expr, tolerance)
 
-    return expr
-#------------------- end of old trigsimp routines --------------------
+    # SymPy's default tolerance for Rationals is 15; other numbers may have
+    # lower tolerances set, so use them to pick the largest tolerance if None
+    # was given
+    if tolerance is None:
+        tolerance = 10**-min([15] +
+             [mpmath.libmp.libmpf.prec_to_dps(n._prec)
+             for n in expr.atoms(Float)])
+    # XXX should prec be set independent of tolerance or should it be computed
+    # from tolerance?
+    prec = 30
+    bprec = int(prec*3.33)
+
+    constants_dict = {}
+    for constant in constants:
+        constant = sympify(constant)
+        v = constant.evalf(prec)
+        if not v.is_Float:
+            raise ValueError("constants must be real-valued")
+        constants_dict[str(constant)] = v._to_mpmath(bprec)
+
+    exprval = expr.evalf(prec, chop=True)
+    re, im = exprval.as_real_imag()
+
+    # safety check to make sure that this evaluated to a number
+    if not (re.is_Number and im.is_Number):
+        return expr
+
+    def nsimplify_real(x):
+        orig = mpmath.mp.dps
+        xv = x._to_mpmath(bprec)
+        try:
+            # We'll be happy with low precision if a simple fraction
+            if not (tolerance or full):
+                mpmath.mp.dps = 15
+                rat = mpmath.findpoly(xv, 1)
+                if rat is not None:
+                    return Rational(-int(rat[1]), int(rat[0]))
+            mpmath.mp.dps = prec
+            newexpr = mpmath.identify(xv, constants=constants_dict,
+                tol=tolerance, full=full)
+            if not newexpr:
+                raise ValueError
+            if full:
+                newexpr = newexpr[0]
+            expr = sympify(newexpr)
+            if x and not expr:  # don't let x become 0
+                raise ValueError
+            if expr.is_finite is False and not xv in [mpmath.inf, mpmath.ninf]:
+                raise ValueError
+            return expr
+        finally:
+            # even though there are returns above, this is executed
+            # before leaving
+            mpmath.mp.dps = orig
+    try:
+        if re:
+            re = nsimplify_real(re)
+        if im:
+            im = nsimplify_real(im)
+    except ValueError:
+        if rational is None:
+            return _real_to_rational(expr)
+        return expr
+
+    rv = re + im*S.ImaginaryUnit
+    # if there was a change or rational is explicitly not wanted
+    # return the value, else return the Rational representation
+    if rv != expr or rational is False:
+        return rv
+    return _real_to_rational(expr)
+
+
+def _real_to_rational(expr, tolerance=None):
+    """
+    Replace all reals in expr with rationals.
+
+    >>> from sympy import nsimplify
+    >>> from sympy.abc import x
+
+    >>> nsimplify(.76 + .1*x**.5, rational=True)
+    sqrt(x)/10 + 19/25
+
+    """
+    p = expr
+    reps = {}
+    reduce_num = None
+    if tolerance is not None and tolerance < 1:
+        reduce_num = ceiling(1/tolerance)
+    for float in p.atoms(Float):
+        key = float
+        if reduce_num is not None:
+            r = Rational(float).limit_denominator(reduce_num)
+        elif (tolerance is not None and tolerance >= 1 and
+                float.is_Integer is False):
+            r = Rational(tolerance*round(float/tolerance)
+                ).limit_denominator(int(tolerance))
+        else:
+            r = nsimplify(float, rational=False)
+            # e.g. log(3).n() -> log(3) instead of a Rational
+            if float and not r:
+                r = Rational(float)
+            elif not r.is_Rational:
+                if float < 0:
+                    float = -float
+                    d = Pow(10, int((mpmath.log(float)/mpmath.log(10))))
+                    r = -Rational(str(float/d))*d
+                elif float > 0:
+                    d = Pow(10, int((mpmath.log(float)/mpmath.log(10))))
+                    r = Rational(str(float/d))*d
+                else:
+                    r = Integer(0)
+        reps[key] = r
+    return p.subs(reps, simultaneous=True)
