+++ sympy/solvers/diophantine.py
@@ -5,7 +5,7 @@ from sympy import (Poly, igcd, divisors, sign, symbols, S, Integer, Wild, Symbol
     isprime, nextprime, integer_nthroot)
 
 from sympy.core.function import _mexpand
-from sympy.simplify.simplify import rad_rationalize
+from sympy.simplify.radsimp import rad_rationalize
 from sympy.utilities import default_sort_key, numbered_symbols
 from sympy.core.numbers import igcdex
 from sympy.ntheory.residue_ntheory import sqrt_mod
